+++ app/app.html
@@ -26,7 +26,7 @@
         scripts.push(
           (process.env.HOT)
             ? 'http://localhost:' + port + '/dist/renderer.dev.js'
-            : './dist/renderer.prod.js'
+            : './dist/app.renderer.prod.js'
         );
 
         document.write(
