+++ app/main.prod.js
@@ -1,4 +1,4 @@
-module.exports=function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}return n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)n.d(r,o,function(t){return e[t]}.bind(null,o));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s="./app/main.dev.js")}({"./app/about.js":function(e,t,n){const{BrowserWindow:r,remote:o,shell:i}=n("electron"),s=n("path");let a=null;e.exports=function(e){if(null!==a)return a.focus(),a;const t="file://"+s.join(__dirname,"about.html"),n={width:400,height:400,useContentSize:!0,titleBarStyle:"hidden-inset",icon:"../resources/icon.ico",parent:e};return(a=new(r||o.BrowserWindow)(n)).once("closed",()=>{a=null}),a.loadURL(t),a.webContents.on("will-navigate",(e,t)=>{e.preventDefault(),i.openExternal(t)}),a.webContents.on("new-window",(e,t)=>{e.preventDefault(),i.openExternal(t)}),a.webContents.once("dom-ready",()=>{let e;e={icon_path:"../resources/icon.ico",product_name:"BloomRPC",copyright:"LGPL-3.0-or-later",homepage:"https://github.com/uw-labs/bloomrpc",description:"GRPC GUI client",license:"LGPL-3.0-or-later",bug_report_url:"https://github.com/uw-labs/bloomrpc/issues",use_version_info:!0},a.webContents.send("about-window:info",e)}),a.once("ready-to-show",()=>{a.show()}),a.setMenu(null),a.setMenuBarVisibility(!1),a}},"./app/main.dev.js":function(e,t,n){const{app:r,BrowserWindow:o}=n("electron"),i=n("./app/menu.js");let s=null;n("./node_modules/source-map-support/source-map-support.js").install();r.on("window-all-closed",()=>{"darwin"!==process.platform&&r.quit()}),r.on("ready",async()=>{(s=new o({show:!1,width:1324,height:728,backgroundColor:"#f0f2f5"})).loadURL(`file://${__dirname}/app.html`),s.once("ready-to-show",()=>{if(!s)throw new Error('"mainWindow" is not defined');setTimeout(function(){s.show(),s.focus()},150)}),s.on("closed",()=>{s=null}),new i(s).buildMenu()})},"./app/menu.js":function(e,t,n){const{app:r,Menu:o,shell:i}=n("electron"),s=n("./app/about.js"),{join:a}=n("path");e.exports=class{constructor(e){this.mainWindow=e}buildMenu(){const e="darwin"===process.platform?this.buildDarwinTemplate():this.buildDefaultTemplate(),t=o.buildFromTemplate(e);return o.setApplicationMenu(t),t}setupDevelopmentEnvironment(){this.mainWindow.openDevTools(),this.mainWindow.webContents.on("context-menu",(e,t)=>{const{x:n,y:r}=t;o.buildFromTemplate([{label:"Inspect element",click:()=>{this.mainWindow.inspectElement(n,r)}}]).popup(this.mainWindow)})}buildDarwinTemplate(){return[{label:"BloomRPC",submenu:[{label:"About BloomRPC",selector:"orderFrontStandardAboutPanel:"},{type:"separator"},{label:"Services",submenu:[]},{type:"separator"},{label:"Hide BloomRPC",accelerator:"Command+H",selector:"hide:"},{label:"Hide Others",accelerator:"Command+Shift+H",selector:"hideOtherApplications:"},{label:"Show All",selector:"unhideAllApplications:"},{type:"separator"},{label:"Quit",accelerator:"Command+Q",click:()=>{r.quit()}}]},{label:"Edit",submenu:[{label:"Undo",accelerator:"Command+Z",selector:"undo:"},{label:"Redo",accelerator:"Shift+Command+Z",selector:"redo:"},{type:"separator"},{label:"Cut",accelerator:"Command+X",selector:"cut:"},{label:"Copy",accelerator:"Command+C",selector:"copy:"},{label:"Paste",accelerator:"Command+V",selector:"paste:"},{label:"Select All",accelerator:"Command+A",selector:"selectAll:"},{label:"Clear Storage",accelerator:"Ctrl+Command+D",click(){n("./app/storage/index.ts").clearAll()}}]},{label:"View",submenu:[{label:"Toggle Full Screen",accelerator:"Ctrl+Command+F",click:()=>{this.mainWindow.setFullScreen(!this.mainWindow.isFullScreen())}}]},{label:"Window",submenu:[{label:"Minimize",accelerator:"Command+M",selector:"performMiniaturize:"},{label:"Close",accelerator:"Command+W",selector:"performClose:"},{type:"separator"},{label:"Bring All to Front",selector:"arrangeInFront:"}]},{label:"Help",submenu:[{label:"Learn More",click(){i.openExternal("https://github.com/uw-labs/bloomrpc")}},{label:"Documentation",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/blob/master/README.md")}},{label:"Community Discussions",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/issues")}},{label:"Search Issues",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/issues")}}]}]}buildDefaultTemplate(){return[{label:"&File",submenu:[{label:"&Open",accelerator:"Ctrl+O"},{label:"&Close",accelerator:"Ctrl+W",click:()=>{this.mainWindow.close()}},{label:"&Quit",accelerator:"Ctrl+Q",click:r.quit}]},{label:"&View",submenu:[{label:"Toggle &Full Screen",accelerator:"F11",click:()=>{this.mainWindow.setFullScreen(!this.mainWindow.isFullScreen())}}]},{label:"Help",submenu:[{label:"Learn More",click(){i.openExternal("https://github.com/uw-labs/bloomrpc")}},{label:"Documentation",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/blob/master/README.md")}},{label:"Community Discussions",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/issues")}},{label:"Search Issues",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/issues")}},{label:"About BloomRPC",click:()=>{s(this.mainWindow)}}]}]}}},"./app/storage/editor.ts":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});const r=new(n("./node_modules/electron-store/index.js"))({name:"editor"});t.EditorStore=r;const o={URL:"url",PROTOS:"protos",TABS:"tabs",REQUESTS:"requests",INTERACTIVE:"interactive",METADATA:"metadata"};t.storeUrl=function(e){r.set(o.URL,e)},t.getUrl=function(){return r.get(o.URL)},t.storeProtos=function(e){r.set(o.PROTOS,e.map(e=>e.proto.filePath))},t.getProtos=function(){return r.get(o.PROTOS)},t.storeTabs=function(e){r.set(o.TABS,{activeKey:e.activeKey,tabs:e.tabs.map(e=>({methodName:e.methodName,serviceName:e.service.serviceName,protoPath:e.service.proto.filePath}))})},t.getTabs=function(){return r.get(o.TABS)},t.storeRequestInfo=function({id:e,url:t,data:n,inputs:i,metadata:s,interactive:a,tlsCertificate:c}){const u={id:e,url:t,data:i||n,metadata:s,interactive:a,tlsCertificate:c,createdAt:(new Date).toISOString()},l=r.get("requests",[]).filter(t=>t.id!==e);r.set(o.REQUESTS,[...l,u])},t.storeMetadata=function(e){r.set(o.METADATA,e)},t.getMetadata=function(){return r.get(o.METADATA)},t.getRequestInfo=function(e){return r.get(o.REQUESTS,[]).find(t=>t.id===e)},t.deleteRequestInfo=function(e){const t=r.get(o.REQUESTS,[]).filter(t=>t.id!==e);r.set("requests",t)},t.clearEditor=function(){r.clear()}},"./app/storage/importPaths.ts":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});const r=new(n("./node_modules/electron-store/index.js"))({name:"importPaths"}),o={IMPORT_PATH:"paths"};t.storeImportPaths=function(e){r.set(o.IMPORT_PATH,e)},t.getImportPaths=function(){return r.get(o.IMPORT_PATH,[""])},t.clearImportPaths=function(){return r.clear()}},"./app/storage/index.ts":function(e,t,n){"use strict";function r(e){for(var n in e)t.hasOwnProperty(n)||(t[n]=e[n])}Object.defineProperty(t,"__esModule",{value:!0});const o=n("./app/storage/editor.ts"),i=n("./app/storage/importPaths.ts"),s=n("./app/storage/tls.ts");r(n("./app/storage/editor.ts")),r(n("./app/storage/importPaths.ts")),r(n("./app/storage/tls.ts")),t.clearAll=function(){o.clearEditor(),i.clearImportPaths(),s.clearTLS()}},"./app/storage/tls.ts":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});const r=new(n("./node_modules/electron-store/index.js"))({name:"tls"}),o={CERTIFICATES:"certificates"};t.storeTLSList=function(e){r.set(o.CERTIFICATES,e)},t.getTLSList=function(){return r.get(o.CERTIFICATES,[])},t.clearTLS=function(){return r.clear()}},"./node_modules/7zip/index.js":function(e,t,n){var r,o,i=n("path").resolve,s=n("./node_modules/7zip/package.json").bin;e.exports=(r=s,o=function(e){return i(__dirname,e)},Object.keys(r).reduce(function(e,t){return e[t]=o(r[t]),e},{}))},"./node_modules/7zip/package.json":function(e){e.exports={name:"7zip",version:"0.0.6",description:"7zip Windows Package via Node.js",keywords:["7z","7zip","7-zip","windows","install"],repository:"git@github.com:fritx/win-7zip.git",bin:{"7z":"7zip-lite/7z.exe"},main:"index.js",scripts:{test:"mocha"},license:"GNU LGPL"}},"./node_modules/balanced-match/index.js":function(e,t,n){"use strict";function r(e,t,n){e instanceof RegExp&&(e=o(e,n)),t instanceof RegExp&&(t=o(t,n));var r=i(e,t,n);return r&&{start:r[0],end:r[1],pre:n.slice(0,r[0]),body:n.slice(r[0]+e.length,r[1]),post:n.slice(r[1]+t.length)}}function o(e,t){var n=t.match(e);return n?n[0]:null}function i(e,t,n){var r,o,i,s,a,c=n.indexOf(e),u=n.indexOf(t,c+1),l=c;if(c>=0&&u>0){for(r=[],i=n.length;l>=0&&!a;)l==c?(r.push(l),c=n.indexOf(e,l+1)):1==r.length?a=[r.pop(),u]:((o=r.pop())<i&&(i=o,s=u),u=n.indexOf(t,l+1)),l=c<u&&c>=0?c:u;r.length&&(a=[i,s])}return a}e.exports=r,r.range=i},"./node_modules/brace-expansion/index.js":function(e,t,n){var r=n("./node_modules/concat-map/index.js"),o=n("./node_modules/balanced-match/index.js");e.exports=function(e){if(!e)return[];"{}"===e.substr(0,2)&&(e="\\{\\}"+e.substr(2));return function e(t,n){var i=[];var s=o("{","}",t);if(!s||/\$$/.test(s.pre))return[t];var c=/^-?\d+\.\.-?\d+(?:\.\.-?\d+)?$/.test(s.body);var u=/^[a-zA-Z]\.\.[a-zA-Z](?:\.\.-?\d+)?$/.test(s.body);var p=c||u;var g=s.body.indexOf(",")>=0;if(!p&&!g)return s.post.match(/,.*\}/)?(t=s.pre+"{"+s.body+a+s.post,e(t)):[t];var y;if(p)y=s.body.split(/\.\./);else if(1===(y=function e(t){if(!t)return[""];var n=[];var r=o("{","}",t);if(!r)return t.split(",");var i=r.pre;var s=r.body;var a=r.post;var c=i.split(",");c[c.length-1]+="{"+s+"}";var u=e(a);a.length&&(c[c.length-1]+=u.shift(),c.push.apply(c,u));n.push.apply(n,c);return n}(s.body)).length&&1===(y=e(y[0],!1).map(h)).length){var v=s.post.length?e(s.post,!1):[""];return v.map(function(e){return s.pre+y[0]+e})}var _=s.pre;var v=s.post.length?e(s.post,!1):[""];var b;if(p){var w=l(y[0]),E=l(y[1]),S=Math.max(y[0].length,y[1].length),j=3==y.length?Math.abs(l(y[2])):1,O=d,x=E<w;x&&(j*=-1,O=m);var A=y.some(f);b=[];for(var C=w;O(C,E);C+=j){var L;if(u)"\\"===(L=String.fromCharCode(C))&&(L="");else if(L=String(C),A){var k=S-L.length;if(k>0){var R=new Array(k+1).join("0");L=C<0?"-"+R+L.slice(1):R+L}}b.push(L)}}else b=r(y,function(t){return e(t,!1)});for(var M=0;M<b.length;M++)for(var T=0;T<v.length;T++){var P=_+b[M]+v[T];(!n||p||P)&&i.push(P)}return i}(function(e){return e.split("\\\\").join(i).split("\\{").join(s).split("\\}").join(a).split("\\,").join(c).split("\\.").join(u)}(e),!0).map(p)};var i="\0SLASH"+Math.random()+"\0",s="\0OPEN"+Math.random()+"\0",a="\0CLOSE"+Math.random()+"\0",c="\0COMMA"+Math.random()+"\0",u="\0PERIOD"+Math.random()+"\0";function l(e){return parseInt(e,10)==e?parseInt(e,10):e.charCodeAt(0)}function p(e){return e.split(i).join("\\").split(s).join("{").split(a).join("}").split(c).join(",").split(u).join(".")}function h(e){return"{"+e+"}"}function f(e){return/^-?0\d/.test(e)}function d(e,t){return e<=t}function m(e,t){return e>=t}},"./node_modules/buffer-from/index.js":function(e,t){var n=Object.prototype.toString,r="function"==typeof Buffer.alloc&&"function"==typeof Buffer.allocUnsafe&&"function"==typeof Buffer.from;e.exports=function(e,t,o){if("number"==typeof e)throw new TypeError('"value" argument must not be a number');return i=e,"ArrayBuffer"===n.call(i).slice(8,-1)?function(e,t,n){t>>>=0;var o=e.byteLength-t;if(o<0)throw new RangeError("'offset' is out of bounds");if(void 0===n)n=o;else if((n>>>=0)>o)throw new RangeError("'length' is out of bounds");return r?Buffer.from(e.slice(t,t+n)):new Buffer(new Uint8Array(e.slice(t,t+n)))}(e,t,o):"string"==typeof e?function(e,t){if("string"==typeof t&&""!==t||(t="utf8"),!Buffer.isEncoding(t))throw new TypeError('"encoding" must be a valid string encoding');return r?Buffer.from(e,t):new Buffer(e,t)}(e,t):r?Buffer.from(e):new Buffer(e);var i}},"./node_modules/concat-map/index.js":function(e,t){e.exports=function(e,t){for(var r=[],o=0;o<e.length;o++){var i=t(e[o],o);n(i)?r.push.apply(r,i):r.push(i)}return r};var n=Array.isArray||function(e){return"[object Array]"===Object.prototype.toString.call(e)}},"./node_modules/conf/index.js":function(e,t,n){"use strict";(function(e){const t=n("fs"),r=n("path"),o=n("crypto"),i=n("assert"),s=n("events"),a=n("./node_modules/dot-prop/index.js"),c=n("./node_modules/make-dir/index.js"),u=n("./node_modules/pkg-up/index.js"),l=n("./node_modules/env-paths/index.js"),p=n("./node_modules/write-file-atomic/index.js"),h=()=>Object.create(null);delete n.c[__filename];const f=r.dirname(e.parent&&e.parent.filename||".");e.exports=class{constructor(e){const n=u.sync(f);if(!(e=Object.assign({projectName:n&&JSON.parse(t.readFileSync(n,"utf8")).name},e)).projectName&&!e.cwd)throw new Error("Project name could not be inferred. Please specify the `projectName` option.");(e=Object.assign({configName:"config",fileExtension:"json",projectSuffix:"nodejs"},e)).cwd||(e.cwd=l(e.projectName,{suffix:e.projectSuffix}).config),this.events=new s,this.encryptionKey=e.encryptionKey;const o=e.fileExtension?`.${e.fileExtension}`:"";this.path=r.resolve(e.cwd,`${e.configName}${o}`);const a=this.store,c=Object.assign(h(),e.defaults,a);try{i.deepEqual(a,c)}catch(e){this.store=c}}get(e,t){return a.get(this.store,e,t)}set(e,t){if("string"!=typeof e&&"object"!=typeof e)throw new TypeError(`Expected \`key\` to be of type \`string\` or \`object\`, got ${typeof e}`);if("object"!=typeof e&&void 0===t)throw new TypeError("Use `delete()` to clear values");const{store:n}=this;if("object"==typeof e)for(const t of Object.keys(e))a.set(n,t,e[t]);else a.set(n,e,t);this.store=n}has(e){return a.has(this.store,e)}delete(e){const{store:t}=this;a.delete(t,e),this.store=t}clear(){this.store=h()}onDidChange(e,t){if("string"!=typeof e)throw new TypeError(`Expected \`key\` to be of type \`string\`, got ${typeof e}`);if("function"!=typeof t)throw new TypeError(`Expected \`callback\` to be of type \`function\`, got ${typeof t}`);let n=this.get(e);const r=()=>{const r=n,o=this.get(e);try{i.deepEqual(o,r)}catch(e){n=o,t.call(this,o,r)}};return this.events.on("change",r),()=>this.events.removeListener("change",r)}get size(){return Object.keys(this.store).length}get store(){try{let e=t.readFileSync(this.path,this.encryptionKey?null:"utf8");if(this.encryptionKey)try{const t=o.createDecipher("aes-256-cbc",this.encryptionKey);e=Buffer.concat([t.update(e),t.final()])}catch(e){}return Object.assign(h(),JSON.parse(e))}catch(e){if("ENOENT"===e.code)return c.sync(r.dirname(this.path)),h();if("SyntaxError"===e.name)return h();throw e}}set store(e){c.sync(r.dirname(this.path));let t=JSON.stringify(e,null,"\t");if(this.encryptionKey){const e=o.createCipher("aes-256-cbc",this.encryptionKey);t=Buffer.concat([e.update(Buffer.from(t)),e.final()])}p.sync(this.path,t),this.events.emit("change")}*[Symbol.iterator](){const{store:e}=this;for(const t of Object.keys(e))yield[t,e[t]]}}}).call(this,n("./node_modules/webpack/buildin/module.js")(e))},"./node_modules/cross-unzip/index.js":function(e,t,n){"use strict";var r=n("child_process").spawn,o=Array.prototype.slice,i="win32"===process.platform?function(e,t,r){s(n("./node_modules/7zip/index.js")["7z"],["x",e,"-y","-o"+t],r)}:function(e,t,n){s("unzip",["-o",e,"-d",t],n)};function s(e,t,n){var i,s;i=n,s=!1,n=function(){s||(s=!0,i.apply(this,o.call(arguments)))};var a=r(e,t,{stdio:"ignore"});a.on("error",function(e){n(e)}),a.on("exit",function(e){n(e?new Error("Exited with code "+e):null)})}i.unzip=i,e.exports=i},"./node_modules/dot-prop/index.js":function(e,t,n){"use strict";const r=n("./node_modules/is-obj/index.js");function o(e){const t=e.split("."),n=[];for(let e=0;e<t.length;e++){let r=t[e];for(;"\\"===r[r.length-1]&&void 0!==t[e+1];)r=r.slice(0,-1)+".",r+=t[++e];n.push(r)}return n}e.exports={get(e,t,n){if(!r(e)||"string"!=typeof t)return void 0===n?e:n;const i=o(t);for(let t=0;t<i.length;t++){if(!Object.prototype.propertyIsEnumerable.call(e,i[t]))return n;if(null==(e=e[i[t]])){if(t!==i.length-1)return n;break}}return e},set(e,t,n){if(!r(e)||"string"!=typeof t)return e;const i=e,s=o(t);for(let t=0;t<s.length;t++){const o=s[t];r(e[o])||(e[o]={}),t===s.length-1&&(e[o]=n),e=e[o]}return i},delete(e,t){if(!r(e)||"string"!=typeof t)return;const n=o(t);for(let t=0;t<n.length;t++){const o=n[t];if(t===n.length-1)return void delete e[o];if(e=e[o],!r(e))return}},has(e,t){if(!r(e)||"string"!=typeof t)return!1;const n=o(t);for(let t=0;t<n.length;t++){if(!r(e))return!1;if(!(n[t]in e))return!1;e=e[n[t]]}return!0}}},"./node_modules/electron-devtools-installer/dist/downloadChromeExtension.js":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});var r=c(n("fs")),o=c(n("path")),i=c(n("./node_modules/rimraf/rimraf.js")),s=c(n("./node_modules/cross-unzip/index.js")),a=n("./node_modules/electron-devtools-installer/dist/utils.js");function c(e){return e&&e.__esModule?e:{default:e}}t.default=function e(t,n){var c=arguments.length>2&&void 0!==arguments[2]?arguments[2]:5,u=(0,a.getPath)();r.default.existsSync(u)||r.default.mkdirSync(u);var l=o.default.resolve(u+"/"+t);return new Promise(function(u,p){if(!r.default.existsSync(l)||n){r.default.existsSync(l)&&i.default.sync(l);var h="https://clients2.google.com/service/update2/crx?response=redirect&x=id%3D"+t+"%26uc&prodversion=32",f=o.default.resolve(l+".crx");(0,a.downloadFile)(h,f).then(function(){(0,s.default)(f,l,function(e){if(e&&!r.default.existsSync(o.default.resolve(l,"manifest.json")))return p(e);(0,a.changePermissions)(l,755),u(l)})}).catch(function(r){if(console.log("Failed to fetch extension, trying "+(c-1)+" more times"),c<=1)return p(r);setTimeout(function(){e(t,n,c-1).then(u).catch(p)},200)})}else u(l)})}},"./node_modules/electron-devtools-installer/dist/index.js":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.MOBX_DEVTOOLS=t.APOLLO_DEVELOPER_TOOLS=t.CYCLEJS_DEVTOOL=t.REACT_PERF=t.REDUX_DEVTOOLS=t.VUEJS_DEVTOOLS=t.ANGULARJS_BATARANG=t.JQUERY_DEBUGGER=t.BACKBONE_DEBUGGER=t.REACT_DEVELOPER_TOOLS=t.EMBER_INSPECTOR=void 0;var r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},o=n("electron"),i=p(o),s=p(n("fs")),a=p(n("path")),c=p(n("./node_modules/semver/semver.js")),u=p(n("./node_modules/electron-devtools-installer/dist/downloadChromeExtension.js")),l=n("./node_modules/electron-devtools-installer/dist/utils.js");function p(e){return e&&e.__esModule?e:{default:e}}var h=(o.remote||i.default).BrowserWindow,f={},d=a.default.resolve((0,l.getPath)(),"IDMap.json");if(s.default.existsSync(d))try{f=JSON.parse(s.default.readFileSync(d,"utf8"))}catch(e){console.error("electron-devtools-installer: Invalid JSON present in the IDMap file")}t.default=function e(t){var n=arguments.length>1&&void 0!==arguments[1]&&arguments[1];if(Array.isArray(t))return Promise.all(t.map(function(t){return e(t,n)}));var o=void 0;if("object"===(void 0===t?"undefined":r(t))&&t.id){o=t.id;var i=process.versions.electron.split("-")[0];if(!c.default.satisfies(i,t.electron))return Promise.reject(new Error("Version of Electron: "+i+" does not match required range "+t.electron+" for extension "+o))}else{if("string"!=typeof t)return Promise.reject(new Error('Invalid extensionReference passed in: "'+t+'"'));o=t}var a=f[o],l=a&&h.getDevToolsExtensions&&h.getDevToolsExtensions()[a];return!n&&l?Promise.resolve(f[o]):(0,u.default)(o,n).then(function(e){l&&h.removeDevToolsExtension(a);var t,n,r,i=h.addDevToolsExtension(e);return s.default.writeFileSync(d,JSON.stringify(Object.assign(f,(r=i,(n=o)in(t={})?Object.defineProperty(t,n,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[n]=r,t)))),Promise.resolve(i)})};t.EMBER_INSPECTOR={id:"bmdblncegkenkacieihfhpjfppoconhi",electron:">=1.2.1"},t.REACT_DEVELOPER_TOOLS={id:"fmkadmapgofadopljbjfkapdkoienihi",electron:">=1.2.1"},t.BACKBONE_DEBUGGER={id:"bhljhndlimiafopmmhjlgfpnnchjjbhd",electron:">=1.2.1"},t.JQUERY_DEBUGGER={id:"dbhhnnnpaeobfddmlalhnehgclcmjimi",electron:">=1.2.1"},t.ANGULARJS_BATARANG={id:"ighdmehidhipcmcojjgiloacoafjmpfk",electron:">=1.2.1"},t.VUEJS_DEVTOOLS={id:"nhdogjmejiglipccpnnnanhbledajbpd",electron:">=1.2.1"},t.REDUX_DEVTOOLS={id:"lmhkpmbekcpmknklioeibfkpmmfibljd",electron:">=1.2.1"},t.REACT_PERF={id:"hacmcodfllhbnekmghgdlplbdnahmhmm",electron:">=1.2.6"},t.CYCLEJS_DEVTOOL={id:"dfgplfmhhmdekalbpejekgfegkonjpfp",electron:">=1.2.1"},t.APOLLO_DEVELOPER_TOOLS={id:"jdkknkkbebbapilgoeccciglkfbmbnfm",electron:">=1.2.1"},t.MOBX_DEVTOOLS={id:"pfgnfdagidkfgccljigdamigbcnndkod",electron:">=1.2.1"}},"./node_modules/electron-devtools-installer/dist/utils.js":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.changePermissions=t.downloadFile=t.getPath=void 0;var r=n("electron"),o=c(r),i=c(n("fs")),s=c(n("path")),a=c(n("https"));function c(e){return e&&e.__esModule?e:{default:e}}t.getPath=function(){var e=(r.remote||o.default).app.getPath("userData");return s.default.resolve(e+"/extensions")};var u=(r.remote||o.default).net,l=u?u.request:a.default.get;t.downloadFile=function e(t,n){return new Promise(function(r,o){var s=l(t);s.on("response",function(t){if(t.statusCode>=300&&t.statusCode<400&&t.headers.location)return e(t.headers.location,n).then(r).catch(o);t.pipe(i.default.createWriteStream(n)).on("close",r)}),s.on("error",o),s.end()})},t.changePermissions=function e(t,n){i.default.readdirSync(t).forEach(function(r){var o=s.default.join(t,r);i.default.chmodSync(o,parseInt(n,8)),i.default.statSync(o).isDirectory()&&e(o,n)})}},"./node_modules/electron-store/index.js":function(e,t,n){"use strict";const r=n("path"),o=n("electron"),i=n("./node_modules/conf/index.js");e.exports=class extends i{constructor(e){const t=(o.app||o.remote.app).getPath("userData");(e=Object.assign({name:"config"},e)).cwd?e.cwd=r.isAbsolute(e.cwd)?e.cwd:r.join(t,e.cwd):e.cwd=t,e.configName=e.name,delete e.name,super(e)}openInEditor(){o.shell.openItem(this.path)}}},"./node_modules/env-paths/index.js":function(e,t,n){"use strict";const r=n("path"),o=n("os"),i=o.homedir(),s=o.tmpdir(),a=process.env;e.exports=((e,t)=>{if("string"!=typeof e)throw new TypeError(`Expected string, got ${typeof e}`);return(t=Object.assign({suffix:"nodejs"},t)).suffix&&(e+=`-${t.suffix}`),"darwin"===process.platform?(e=>{const t=r.join(i,"Library");return{data:r.join(t,"Application Support",e),config:r.join(t,"Preferences",e),cache:r.join(t,"Caches",e),log:r.join(t,"Logs",e),temp:r.join(s,e)}})(e):"win32"===process.platform?(e=>{const t=a.LOCALAPPDATA||r.join(i,"AppData","Local");return{data:r.join(t,e,"Data"),config:r.join(t,e,"Config"),cache:r.join(t,e,"Cache"),log:r.join(t,e,"Log"),temp:r.join(s,e)}})(e):(e=>{const t=r.basename(i);return{data:r.join(a.XDG_DATA_HOME||r.join(i,".local","share"),e),config:r.join(a.XDG_CONFIG_HOME||r.join(i,".config"),e),cache:r.join(a.XDG_CACHE_HOME||r.join(i,".cache"),e),log:r.join(a.XDG_STATE_HOME||r.join(i,".local","state"),e),temp:r.join(s,t,e)}})(e)})},"./node_modules/fs.realpath/index.js":function(e,t,n){e.exports=l,l.realpath=l,l.sync=p,l.realpathSync=p,l.monkeypatch=function(){r.realpath=l,r.realpathSync=p},l.unmonkeypatch=function(){r.realpath=o,r.realpathSync=i};var r=n("fs"),o=r.realpath,i=r.realpathSync,s=process.version,a=/^v[0-5]\./.test(s),c=n("./node_modules/fs.realpath/old.js");function u(e){return e&&"realpath"===e.syscall&&("ELOOP"===e.code||"ENOMEM"===e.code||"ENAMETOOLONG"===e.code)}function l(e,t,n){if(a)return o(e,t,n);"function"==typeof t&&(n=t,t=null),o(e,t,function(r,o){u(r)?c.realpath(e,t,n):n(r,o)})}function p(e,t){if(a)return i(e,t);try{return i(e,t)}catch(n){if(u(n))return c.realpathSync(e,t);throw n}}},"./node_modules/fs.realpath/old.js":function(e,t,n){var r=n("path"),o="win32"===process.platform,i=n("fs"),s=process.env.NODE_DEBUG&&/fs/.test(process.env.NODE_DEBUG);function a(e){return"function"==typeof e?e:function(){var e;if(s){var t=new Error;e=function(e){e&&(t.message=e.message,n(e=t))}}else e=n;return e;function n(e){if(e){if(process.throwDeprecation)throw e;if(!process.noDeprecation){var t="fs: missing callback "+(e.stack||e.message);process.traceDeprecation?console.trace(t):console.error(t)}}}}()}r.normalize;if(o)var c=/(.*?)(?:[\/\\]+|$)/g;else c=/(.*?)(?:[\/]+|$)/g;if(o)var u=/^(?:[a-zA-Z]:|[\\\/]{2}[^\\\/]+[\\\/][^\\\/]+)?[\\\/]*/;else u=/^[\/]*/;t.realpathSync=function(e,t){if(e=r.resolve(e),t&&Object.prototype.hasOwnProperty.call(t,e))return t[e];var n,s,a,l,p=e,h={},f={};function d(){var t=u.exec(e);n=t[0].length,s=t[0],a=t[0],l="",o&&!f[a]&&(i.lstatSync(a),f[a]=!0)}for(d();n<e.length;){c.lastIndex=n;var m=c.exec(e);if(l=s,s+=m[0],a=l+m[1],n=c.lastIndex,!(f[a]||t&&t[a]===a)){var g;if(t&&Object.prototype.hasOwnProperty.call(t,a))g=t[a];else{var y=i.lstatSync(a);if(!y.isSymbolicLink()){f[a]=!0,t&&(t[a]=a);continue}var v=null;if(!o){var _=y.dev.toString(32)+":"+y.ino.toString(32);h.hasOwnProperty(_)&&(v=h[_])}null===v&&(i.statSync(a),v=i.readlinkSync(a)),g=r.resolve(l,v),t&&(t[a]=g),o||(h[_]=v)}e=r.resolve(g,e.slice(n)),d()}}return t&&(t[p]=e),e},t.realpath=function(e,t,n){if("function"!=typeof n&&(n=a(t),t=null),e=r.resolve(e),t&&Object.prototype.hasOwnProperty.call(t,e))return process.nextTick(n.bind(null,null,t[e]));var s,l,p,h,f=e,d={},m={};function g(){var t=u.exec(e);s=t[0].length,l=t[0],p=t[0],h="",o&&!m[p]?i.lstat(p,function(e){if(e)return n(e);m[p]=!0,y()}):process.nextTick(y)}function y(){if(s>=e.length)return t&&(t[f]=e),n(null,e);c.lastIndex=s;var r=c.exec(e);return h=l,l+=r[0],p=h+r[1],s=c.lastIndex,m[p]||t&&t[p]===p?process.nextTick(y):t&&Object.prototype.hasOwnProperty.call(t,p)?b(t[p]):i.lstat(p,v)}function v(e,r){if(e)return n(e);if(!r.isSymbolicLink())return m[p]=!0,t&&(t[p]=p),process.nextTick(y);if(!o){var s=r.dev.toString(32)+":"+r.ino.toString(32);if(d.hasOwnProperty(s))return _(null,d[s],p)}i.stat(p,function(e){if(e)return n(e);i.readlink(p,function(e,t){o||(d[s]=t),_(e,t)})})}function _(e,o,i){if(e)return n(e);var s=r.resolve(h,o);t&&(t[i]=s),b(s)}function b(t){e=r.resolve(t,e.slice(s)),g()}g()}},"./node_modules/glob/common.js":function(e,t,n){function r(e,t){return Object.prototype.hasOwnProperty.call(e,t)}t.alphasort=u,t.alphasorti=c,t.setopts=function(e,t,n){n||(n={});if(n.matchBase&&-1===t.indexOf("/")){if(n.noglobstar)throw new Error("base matching requires globstar");t="**/"+t}e.silent=!!n.silent,e.pattern=t,e.strict=!1!==n.strict,e.realpath=!!n.realpath,e.realpathCache=n.realpathCache||Object.create(null),e.follow=!!n.follow,e.dot=!!n.dot,e.mark=!!n.mark,e.nodir=!!n.nodir,e.nodir&&(e.mark=!0);e.sync=!!n.sync,e.nounique=!!n.nounique,e.nonull=!!n.nonull,e.nosort=!!n.nosort,e.nocase=!!n.nocase,e.stat=!!n.stat,e.noprocess=!!n.noprocess,e.absolute=!!n.absolute,e.maxLength=n.maxLength||1/0,e.cache=n.cache||Object.create(null),e.statCache=n.statCache||Object.create(null),e.symlinks=n.symlinks||Object.create(null),function(e,t){e.ignore=t.ignore||[],Array.isArray(e.ignore)||(e.ignore=[e.ignore]);e.ignore.length&&(e.ignore=e.ignore.map(l))}(e,n),e.changedCwd=!1;var i=process.cwd();r(n,"cwd")?(e.cwd=o.resolve(n.cwd),e.changedCwd=e.cwd!==i):e.cwd=i;e.root=n.root||o.resolve(e.cwd,"/"),e.root=o.resolve(e.root),"win32"===process.platform&&(e.root=e.root.replace(/\\/g,"/"));e.cwdAbs=s(e.cwd)?e.cwd:p(e,e.cwd),"win32"===process.platform&&(e.cwdAbs=e.cwdAbs.replace(/\\/g,"/"));e.nomount=!!n.nomount,n.nonegate=!0,n.nocomment=!0,e.minimatch=new a(t,n),e.options=e.minimatch.options},t.ownProp=r,t.makeAbs=p,t.finish=function(e){for(var t=e.nounique,n=t?[]:Object.create(null),r=0,o=e.matches.length;r<o;r++){var i=e.matches[r];if(i&&0!==Object.keys(i).length){var s=Object.keys(i);t?n.push.apply(n,s):s.forEach(function(e){n[e]=!0})}else if(e.nonull){var a=e.minimatch.globSet[r];t?n.push(a):n[a]=!0}}t||(n=Object.keys(n));e.nosort||(n=n.sort(e.nocase?c:u));if(e.mark){for(var r=0;r<n.length;r++)n[r]=e._mark(n[r]);e.nodir&&(n=n.filter(function(t){var n=!/\/$/.test(t),r=e.cache[t]||e.cache[p(e,t)];return n&&r&&(n="DIR"!==r&&!Array.isArray(r)),n}))}e.ignore.length&&(n=n.filter(function(t){return!h(e,t)}));e.found=n},t.mark=function(e,t){var n=p(e,t),r=e.cache[n],o=t;if(r){var i="DIR"===r||Array.isArray(r),s="/"===t.slice(-1);if(i&&!s?o+="/":!i&&s&&(o=o.slice(0,-1)),o!==t){var a=p(e,o);e.statCache[a]=e.statCache[n],e.cache[a]=e.cache[n]}}return o},t.isIgnored=h,t.childrenIgnored=function(e,t){return!!e.ignore.length&&e.ignore.some(function(e){return!(!e.gmatcher||!e.gmatcher.match(t))})};var o=n("path"),i=n("./node_modules/minimatch/minimatch.js"),s=n("./node_modules/path-is-absolute/index.js"),a=i.Minimatch;function c(e,t){return e.toLowerCase().localeCompare(t.toLowerCase())}function u(e,t){return e.localeCompare(t)}function l(e){var t=null;if("/**"===e.slice(-3)){var n=e.replace(/(\/\*\*)+$/,"");t=new a(n,{dot:!0})}return{matcher:new a(e,{dot:!0}),gmatcher:t}}function p(e,t){var n=t;return n="/"===t.charAt(0)?o.join(e.root,t):s(t)||""===t?t:e.changedCwd?o.resolve(e.cwd,t):o.resolve(t),"win32"===process.platform&&(n=n.replace(/\\/g,"/")),n}function h(e,t){return!!e.ignore.length&&e.ignore.some(function(e){return e.matcher.match(t)||!(!e.gmatcher||!e.gmatcher.match(t))})}},"./node_modules/glob/glob.js":function(e,t,n){e.exports=_;var r=n("fs"),o=n("./node_modules/fs.realpath/index.js"),i=n("./node_modules/minimatch/minimatch.js"),s=(i.Minimatch,n("./node_modules/inherits/inherits.js")),a=n("events").EventEmitter,c=n("path"),u=n("assert"),l=n("./node_modules/path-is-absolute/index.js"),p=n("./node_modules/glob/sync.js"),h=n("./node_modules/glob/common.js"),f=(h.alphasort,h.alphasorti,h.setopts),d=h.ownProp,m=n("./node_modules/inflight/inflight.js"),g=(n("util"),h.childrenIgnored),y=h.isIgnored,v=n("./node_modules/once/once.js");function _(e,t,n){if("function"==typeof t&&(n=t,t={}),t||(t={}),t.sync){if(n)throw new TypeError("callback provided to sync glob");return p(e,t)}return new w(e,t,n)}_.sync=p;var b=_.GlobSync=p.GlobSync;function w(e,t,n){if("function"==typeof t&&(n=t,t=null),t&&t.sync){if(n)throw new TypeError("callback provided to sync glob");return new b(e,t)}if(!(this instanceof w))return new w(e,t,n);f(this,e,t),this._didRealPath=!1;var r=this.minimatch.set.length;this.matches=new Array(r),"function"==typeof n&&(n=v(n),this.on("error",n),this.on("end",function(e){n(null,e)}));var o=this;if(this._processing=0,this._emitQueue=[],this._processQueue=[],this.paused=!1,this.noprocess)return this;if(0===r)return a();for(var i=!0,s=0;s<r;s++)this._process(this.minimatch.set[s],s,!1,a);function a(){--o._processing,o._processing<=0&&(i?process.nextTick(function(){o._finish()}):o._finish())}i=!1}_.glob=_,_.hasMagic=function(e,t){var n=function(e,t){if(null===t||"object"!=typeof t)return e;for(var n=Object.keys(t),r=n.length;r--;)e[n[r]]=t[n[r]];return e}({},t);n.noprocess=!0;var r=new w(e,n).minimatch.set;if(!e)return!1;if(r.length>1)return!0;for(var o=0;o<r[0].length;o++)if("string"!=typeof r[0][o])return!0;return!1},_.Glob=w,s(w,a),w.prototype._finish=function(){if(u(this instanceof w),!this.aborted){if(this.realpath&&!this._didRealpath)return this._realpath();h.finish(this),this.emit("end",this.found)}},w.prototype._realpath=function(){if(!this._didRealpath){this._didRealpath=!0;var e=this.matches.length;if(0===e)return this._finish();for(var t=this,n=0;n<this.matches.length;n++)this._realpathSet(n,r)}function r(){0==--e&&t._finish()}},w.prototype._realpathSet=function(e,t){var n=this.matches[e];if(!n)return t();var r=Object.keys(n),i=this,s=r.length;if(0===s)return t();var a=this.matches[e]=Object.create(null);r.forEach(function(n,r){n=i._makeAbs(n),o.realpath(n,i.realpathCache,function(r,o){r?"stat"===r.syscall?a[n]=!0:i.emit("error",r):a[o]=!0,0==--s&&(i.matches[e]=a,t())})})},w.prototype._mark=function(e){return h.mark(this,e)},w.prototype._makeAbs=function(e){return h.makeAbs(this,e)},w.prototype.abort=function(){this.aborted=!0,this.emit("abort")},w.prototype.pause=function(){this.paused||(this.paused=!0,this.emit("pause"))},w.prototype.resume=function(){if(this.paused){if(this.emit("resume"),this.paused=!1,this._emitQueue.length){var e=this._emitQueue.slice(0);this._emitQueue.length=0;for(var t=0;t<e.length;t++){var n=e[t];this._emitMatch(n[0],n[1])}}if(this._processQueue.length){var r=this._processQueue.slice(0);this._processQueue.length=0;for(t=0;t<r.length;t++){var o=r[t];this._processing--,this._process(o[0],o[1],o[2],o[3])}}}},w.prototype._process=function(e,t,n,r){if(u(this instanceof w),u("function"==typeof r),!this.aborted)if(this._processing++,this.paused)this._processQueue.push([e,t,n,r]);else{for(var o,s=0;"string"==typeof e[s];)s++;switch(s){case e.length:return void this._processSimple(e.join("/"),t,r);case 0:o=null;break;default:o=e.slice(0,s).join("/")}var a,c=e.slice(s);null===o?a=".":l(o)||l(e.join("/"))?(o&&l(o)||(o="/"+o),a=o):a=o;var p=this._makeAbs(a);if(g(this,a))return r();c[0]===i.GLOBSTAR?this._processGlobStar(o,a,p,c,t,n,r):this._processReaddir(o,a,p,c,t,n,r)}},w.prototype._processReaddir=function(e,t,n,r,o,i,s){var a=this;this._readdir(n,i,function(c,u){return a._processReaddir2(e,t,n,r,o,i,u,s)})},w.prototype._processReaddir2=function(e,t,n,r,o,i,s,a){if(!s)return a();for(var u=r[0],l=!!this.minimatch.negate,p=u._glob,h=this.dot||"."===p.charAt(0),f=[],d=0;d<s.length;d++){if("."!==(g=s[d]).charAt(0)||h)(l&&!e?!g.match(u):g.match(u))&&f.push(g)}var m=f.length;if(0===m)return a();if(1===r.length&&!this.mark&&!this.stat){this.matches[o]||(this.matches[o]=Object.create(null));for(d=0;d<m;d++){var g=f[d];e&&(g="/"!==e?e+"/"+g:e+g),"/"!==g.charAt(0)||this.nomount||(g=c.join(this.root,g)),this._emitMatch(o,g)}return a()}r.shift();for(d=0;d<m;d++){g=f[d];e&&(g="/"!==e?e+"/"+g:e+g),this._process([g].concat(r),o,i,a)}a()},w.prototype._emitMatch=function(e,t){if(!this.aborted&&!y(this,t))if(this.paused)this._emitQueue.push([e,t]);else{var n=l(t)?t:this._makeAbs(t);if(this.mark&&(t=this._mark(t)),this.absolute&&(t=n),!this.matches[e][t]){if(this.nodir){var r=this.cache[n];if("DIR"===r||Array.isArray(r))return}this.matches[e][t]=!0;var o=this.statCache[n];o&&this.emit("stat",t,o),this.emit("match",t)}}},w.prototype._readdirInGlobStar=function(e,t){if(!this.aborted){if(this.follow)return this._readdir(e,!1,t);var n=this,o=m("lstat\0"+e,function(r,o){if(r&&"ENOENT"===r.code)return t();var i=o&&o.isSymbolicLink();n.symlinks[e]=i,i||!o||o.isDirectory()?n._readdir(e,!1,t):(n.cache[e]="FILE",t())});o&&r.lstat(e,o)}},w.prototype._readdir=function(e,t,n){if(!this.aborted&&(n=m("readdir\0"+e+"\0"+t,n))){if(t&&!d(this.symlinks,e))return this._readdirInGlobStar(e,n);if(d(this.cache,e)){var o=this.cache[e];if(!o||"FILE"===o)return n();if(Array.isArray(o))return n(null,o)}r.readdir(e,function(e,t,n){return function(r,o){r?e._readdirError(t,r,n):e._readdirEntries(t,o,n)}}(this,e,n))}},w.prototype._readdirEntries=function(e,t,n){if(!this.aborted){if(!this.mark&&!this.stat)for(var r=0;r<t.length;r++){var o=t[r];o="/"===e?e+o:e+"/"+o,this.cache[o]=!0}return this.cache[e]=t,n(null,t)}},w.prototype._readdirError=function(e,t,n){if(!this.aborted){switch(t.code){case"ENOTSUP":case"ENOTDIR":var r=this._makeAbs(e);if(this.cache[r]="FILE",r===this.cwdAbs){var o=new Error(t.code+" invalid cwd "+this.cwd);o.path=this.cwd,o.code=t.code,this.emit("error",o),this.abort()}break;case"ENOENT":case"ELOOP":case"ENAMETOOLONG":case"UNKNOWN":this.cache[this._makeAbs(e)]=!1;break;default:this.cache[this._makeAbs(e)]=!1,this.strict&&(this.emit("error",t),this.abort()),this.silent||console.error("glob error",t)}return n()}},w.prototype._processGlobStar=function(e,t,n,r,o,i,s){var a=this;this._readdir(n,i,function(c,u){a._processGlobStar2(e,t,n,r,o,i,u,s)})},w.prototype._processGlobStar2=function(e,t,n,r,o,i,s,a){if(!s)return a();var c=r.slice(1),u=e?[e]:[],l=u.concat(c);this._process(l,o,!1,a);var p=this.symlinks[n],h=s.length;if(p&&i)return a();for(var f=0;f<h;f++){if("."!==s[f].charAt(0)||this.dot){var d=u.concat(s[f],c);this._process(d,o,!0,a);var m=u.concat(s[f],r);this._process(m,o,!0,a)}}a()},w.prototype._processSimple=function(e,t,n){var r=this;this._stat(e,function(o,i){r._processSimple2(e,t,o,i,n)})},w.prototype._processSimple2=function(e,t,n,r,o){if(this.matches[t]||(this.matches[t]=Object.create(null)),!r)return o();if(e&&l(e)&&!this.nomount){var i=/[\/\\]$/.test(e);"/"===e.charAt(0)?e=c.join(this.root,e):(e=c.resolve(this.root,e),i&&(e+="/"))}"win32"===process.platform&&(e=e.replace(/\\/g,"/")),this._emitMatch(t,e),o()},w.prototype._stat=function(e,t){var n=this._makeAbs(e),o="/"===e.slice(-1);if(e.length>this.maxLength)return t();if(!this.stat&&d(this.cache,n)){var i=this.cache[n];if(Array.isArray(i)&&(i="DIR"),!o||"DIR"===i)return t(null,i);if(o&&"FILE"===i)return t()}var s=this.statCache[n];if(void 0!==s){if(!1===s)return t(null,s);var a=s.isDirectory()?"DIR":"FILE";return o&&"FILE"===a?t():t(null,a,s)}var c=this,u=m("stat\0"+n,function(o,i){if(i&&i.isSymbolicLink())return r.stat(n,function(r,o){r?c._stat2(e,n,null,i,t):c._stat2(e,n,r,o,t)});c._stat2(e,n,o,i,t)});u&&r.lstat(n,u)},w.prototype._stat2=function(e,t,n,r,o){if(n&&("ENOENT"===n.code||"ENOTDIR"===n.code))return this.statCache[t]=!1,o();var i="/"===e.slice(-1);if(this.statCache[t]=r,"/"===t.slice(-1)&&r&&!r.isDirectory())return o(null,!1,r);var s=!0;return r&&(s=r.isDirectory()?"DIR":"FILE"),this.cache[t]=this.cache[t]||s,i&&"FILE"===s?o():o(null,s,r)}},"./node_modules/glob/sync.js":function(e,t,n){e.exports=d,d.GlobSync=m;var r=n("fs"),o=n("./node_modules/fs.realpath/index.js"),i=n("./node_modules/minimatch/minimatch.js"),s=(i.Minimatch,n("./node_modules/glob/glob.js").Glob,n("util"),n("path")),a=n("assert"),c=n("./node_modules/path-is-absolute/index.js"),u=n("./node_modules/glob/common.js"),l=(u.alphasort,u.alphasorti,u.setopts),p=u.ownProp,h=u.childrenIgnored,f=u.isIgnored;function d(e,t){if("function"==typeof t||3===arguments.length)throw new TypeError("callback provided to sync glob\nSee: https://github.com/isaacs/node-glob/issues/167");return new m(e,t).found}function m(e,t){if(!e)throw new Error("must provide pattern");if("function"==typeof t||3===arguments.length)throw new TypeError("callback provided to sync glob\nSee: https://github.com/isaacs/node-glob/issues/167");if(!(this instanceof m))return new m(e,t);if(l(this,e,t),this.noprocess)return this;var n=this.minimatch.set.length;this.matches=new Array(n);for(var r=0;r<n;r++)this._process(this.minimatch.set[r],r,!1);this._finish()}m.prototype._finish=function(){if(a(this instanceof m),this.realpath){var e=this;this.matches.forEach(function(t,n){var r=e.matches[n]=Object.create(null);for(var i in t)try{i=e._makeAbs(i),r[o.realpathSync(i,e.realpathCache)]=!0}catch(t){if("stat"!==t.syscall)throw t;r[e._makeAbs(i)]=!0}})}u.finish(this)},m.prototype._process=function(e,t,n){a(this instanceof m);for(var r,o=0;"string"==typeof e[o];)o++;switch(o){case e.length:return void this._processSimple(e.join("/"),t);case 0:r=null;break;default:r=e.slice(0,o).join("/")}var s,u=e.slice(o);null===r?s=".":c(r)||c(e.join("/"))?(r&&c(r)||(r="/"+r),s=r):s=r;var l=this._makeAbs(s);h(this,s)||(u[0]===i.GLOBSTAR?this._processGlobStar(r,s,l,u,t,n):this._processReaddir(r,s,l,u,t,n))},m.prototype._processReaddir=function(e,t,n,r,o,i){var a=this._readdir(n,i);if(a){for(var c=r[0],u=!!this.minimatch.negate,l=c._glob,p=this.dot||"."===l.charAt(0),h=[],f=0;f<a.length;f++){if("."!==(g=a[f]).charAt(0)||p)(u&&!e?!g.match(c):g.match(c))&&h.push(g)}var d=h.length;if(0!==d)if(1!==r.length||this.mark||this.stat){r.shift();for(f=0;f<d;f++){var m;g=h[f];m=e?[e,g]:[g],this._process(m.concat(r),o,i)}}else{this.matches[o]||(this.matches[o]=Object.create(null));for(var f=0;f<d;f++){var g=h[f];e&&(g="/"!==e.slice(-1)?e+"/"+g:e+g),"/"!==g.charAt(0)||this.nomount||(g=s.join(this.root,g)),this._emitMatch(o,g)}}}},m.prototype._emitMatch=function(e,t){if(!f(this,t)){var n=this._makeAbs(t);if(this.mark&&(t=this._mark(t)),this.absolute&&(t=n),!this.matches[e][t]){if(this.nodir){var r=this.cache[n];if("DIR"===r||Array.isArray(r))return}this.matches[e][t]=!0,this.stat&&this._stat(t)}}},m.prototype._readdirInGlobStar=function(e){if(this.follow)return this._readdir(e,!1);var t,n;try{n=r.lstatSync(e)}catch(e){if("ENOENT"===e.code)return null}var o=n&&n.isSymbolicLink();return this.symlinks[e]=o,o||!n||n.isDirectory()?t=this._readdir(e,!1):this.cache[e]="FILE",t},m.prototype._readdir=function(e,t){if(t&&!p(this.symlinks,e))return this._readdirInGlobStar(e);if(p(this.cache,e)){var n=this.cache[e];if(!n||"FILE"===n)return null;if(Array.isArray(n))return n}try{return this._readdirEntries(e,r.readdirSync(e))}catch(t){return this._readdirError(e,t),null}},m.prototype._readdirEntries=function(e,t){if(!this.mark&&!this.stat)for(var n=0;n<t.length;n++){var r=t[n];r="/"===e?e+r:e+"/"+r,this.cache[r]=!0}return this.cache[e]=t,t},m.prototype._readdirError=function(e,t){switch(t.code){case"ENOTSUP":case"ENOTDIR":var n=this._makeAbs(e);if(this.cache[n]="FILE",n===this.cwdAbs){var r=new Error(t.code+" invalid cwd "+this.cwd);throw r.path=this.cwd,r.code=t.code,r}break;case"ENOENT":case"ELOOP":case"ENAMETOOLONG":case"UNKNOWN":this.cache[this._makeAbs(e)]=!1;break;default:if(this.cache[this._makeAbs(e)]=!1,this.strict)throw t;this.silent||console.error("glob error",t)}},m.prototype._processGlobStar=function(e,t,n,r,o,i){var s=this._readdir(n,i);if(s){var a=r.slice(1),c=e?[e]:[],u=c.concat(a);this._process(u,o,!1);var l=s.length;if(!this.symlinks[n]||!i)for(var p=0;p<l;p++){if("."!==s[p].charAt(0)||this.dot){var h=c.concat(s[p],a);this._process(h,o,!0);var f=c.concat(s[p],r);this._process(f,o,!0)}}}},m.prototype._processSimple=function(e,t){var n=this._stat(e);if(this.matches[t]||(this.matches[t]=Object.create(null)),n){if(e&&c(e)&&!this.nomount){var r=/[\/\\]$/.test(e);"/"===e.charAt(0)?e=s.join(this.root,e):(e=s.resolve(this.root,e),r&&(e+="/"))}"win32"===process.platform&&(e=e.replace(/\\/g,"/")),this._emitMatch(t,e)}},m.prototype._stat=function(e){var t=this._makeAbs(e),n="/"===e.slice(-1);if(e.length>this.maxLength)return!1;if(!this.stat&&p(this.cache,t)){var o=this.cache[t];if(Array.isArray(o)&&(o="DIR"),!n||"DIR"===o)return o;if(n&&"FILE"===o)return!1}var i=this.statCache[t];if(!i){var s;try{s=r.lstatSync(t)}catch(e){if(e&&("ENOENT"===e.code||"ENOTDIR"===e.code))return this.statCache[t]=!1,!1}if(s&&s.isSymbolicLink())try{i=r.statSync(t)}catch(e){i=s}else i=s}this.statCache[t]=i;o=!0;return i&&(o=i.isDirectory()?"DIR":"FILE"),this.cache[t]=this.cache[t]||o,(!n||"FILE"!==o)&&o},m.prototype._mark=function(e){return u.mark(this,e)},m.prototype._makeAbs=function(e){return u.makeAbs(this,e)}},"./node_modules/graceful-fs/clone.js":function(e,t,n){"use strict";e.exports=function(e){if(null===e||"object"!=typeof e)return e;if(e instanceof Object)var t={__proto__:e.__proto__};else var t=Object.create(null);return Object.getOwnPropertyNames(e).forEach(function(n){Object.defineProperty(t,n,Object.getOwnPropertyDescriptor(e,n))}),t}},"./node_modules/graceful-fs/graceful-fs.js":function(e,t,n){var r=n("fs"),o=n("./node_modules/graceful-fs/polyfills.js"),i=n("./node_modules/graceful-fs/legacy-streams.js"),s=n("./node_modules/graceful-fs/clone.js"),a=[],c=n("util");var u,l,p=function(){};function h(e){o(e),e.gracefulify=h,e.FileReadStream=p,e.FileWriteStream=m,e.createReadStream=function(e,t){return new p(e,t)},e.createWriteStream=function(e,t){return new m(e,t)};var t=e.readFile;e.readFile=function(e,n,r){"function"==typeof n&&(r=n,n=null);return function e(n,r,o){return t(n,r,function(t){!t||"EMFILE"!==t.code&&"ENFILE"!==t.code?("function"==typeof o&&o.apply(this,arguments),d()):f([e,[n,r,o]])})}(e,n,r)};var n=e.writeFile;e.writeFile=function(e,t,r,o){"function"==typeof r&&(o=r,r=null);return function e(t,r,o,i){return n(t,r,o,function(n){!n||"EMFILE"!==n.code&&"ENFILE"!==n.code?("function"==typeof i&&i.apply(this,arguments),d()):f([e,[t,r,o,i]])})}(e,t,r,o)};var r=e.appendFile;r&&(e.appendFile=function(e,t,n,o){"function"==typeof n&&(o=n,n=null);return function e(t,n,o,i){return r(t,n,o,function(r){!r||"EMFILE"!==r.code&&"ENFILE"!==r.code?("function"==typeof i&&i.apply(this,arguments),d()):f([e,[t,n,o,i]])})}(e,t,n,o)});var s=e.readdir;function a(t){return s.apply(e,t)}if(e.readdir=function(e,t,n){var r=[e];"function"!=typeof t?r.push(t):n=t;return r.push(function(e,t){t&&t.sort&&t.sort(),!e||"EMFILE"!==e.code&&"ENFILE"!==e.code?("function"==typeof n&&n.apply(this,arguments),d()):f([a,[r]])}),a(r)},"v0.8"===process.version.substr(0,4)){var c=i(e);p=c.ReadStream,m=c.WriteStream}var u=e.ReadStream;u&&(p.prototype=Object.create(u.prototype),p.prototype.open=function(){var e=this;y(e.path,e.flags,e.mode,function(t,n){t?(e.autoClose&&e.destroy(),e.emit("error",t)):(e.fd=n,e.emit("open",n),e.read())})});var l=e.WriteStream;function p(e,t){return this instanceof p?(u.apply(this,arguments),this):p.apply(Object.create(p.prototype),arguments)}function m(e,t){return this instanceof m?(l.apply(this,arguments),this):m.apply(Object.create(m.prototype),arguments)}l&&(m.prototype=Object.create(l.prototype),m.prototype.open=function(){var e=this;y(e.path,e.flags,e.mode,function(t,n){t?(e.destroy(),e.emit("error",t)):(e.fd=n,e.emit("open",n))})}),e.ReadStream=p,e.WriteStream=m;var g=e.open;function y(e,t,n,r){return"function"==typeof n&&(r=n,n=null),function e(t,n,r,o){return g(t,n,r,function(i,s){!i||"EMFILE"!==i.code&&"ENFILE"!==i.code?("function"==typeof o&&o.apply(this,arguments),d()):f([e,[t,n,r,o]])})}(e,t,n,r)}return e.open=y,e}function f(e){p("ENQUEUE",e[0].name,e[1]),a.push(e)}function d(){var e=a.shift();e&&(p("RETRY",e[0].name,e[1]),e[0].apply(null,e[1]))}c.debuglog?p=c.debuglog("gfs4"):/\bgfs4\b/i.test(process.env.NODE_DEBUG||"")&&(p=function(){var e=c.format.apply(c,arguments);e="GFS4: "+e.split(/\n/).join("\nGFS4: "),console.error(e)}),/\bgfs4\b/i.test(process.env.NODE_DEBUG||"")&&process.on("exit",function(){p(a),n("assert").equal(a.length,0)}),e.exports=h(s(r)),process.env.TEST_GRACEFUL_FS_GLOBAL_PATCH&&!r.__patched&&(e.exports=h(r),r.__patched=!0),e.exports.close=(u=r.close,function(e,t){return u.call(r,e,function(e){e||d(),"function"==typeof t&&t.apply(this,arguments)})}),e.exports.closeSync=(l=r.closeSync,function(e){var t=l.apply(r,arguments);return d(),t}),/\bgraceful-fs\b/.test(r.closeSync.toString())||(r.closeSync=e.exports.closeSync,r.close=e.exports.close)},"./node_modules/graceful-fs/legacy-streams.js":function(e,t,n){var r=n("stream").Stream;e.exports=function(e){return{ReadStream:function t(n,o){if(!(this instanceof t))return new t(n,o);r.call(this);var i=this;this.path=n;this.fd=null;this.readable=!0;this.paused=!1;this.flags="r";this.mode=438;this.bufferSize=65536;o=o||{};var s=Object.keys(o);for(var a=0,c=s.length;a<c;a++){var u=s[a];this[u]=o[u]}this.encoding&&this.setEncoding(this.encoding);if(void 0!==this.start){if("number"!=typeof this.start)throw TypeError("start must be a Number");if(void 0===this.end)this.end=1/0;else if("number"!=typeof this.end)throw TypeError("end must be a Number");if(this.start>this.end)throw new Error("start must be <= end");this.pos=this.start}if(null!==this.fd)return void process.nextTick(function(){i._read()});e.open(this.path,this.flags,this.mode,function(e,t){if(e)return i.emit("error",e),void(i.readable=!1);i.fd=t,i.emit("open",t),i._read()})},WriteStream:function t(n,o){if(!(this instanceof t))return new t(n,o);r.call(this);this.path=n;this.fd=null;this.writable=!0;this.flags="w";this.encoding="binary";this.mode=438;this.bytesWritten=0;o=o||{};var i=Object.keys(o);for(var s=0,a=i.length;s<a;s++){var c=i[s];this[c]=o[c]}if(void 0!==this.start){if("number"!=typeof this.start)throw TypeError("start must be a Number");if(this.start<0)throw new Error("start must be >= zero");this.pos=this.start}this.busy=!1;this._queue=[];null===this.fd&&(this._open=e.open,this._queue.push([this._open,this.path,this.flags,this.mode,void 0]),this.flush())}}}},"./node_modules/graceful-fs/polyfills.js":function(e,t,n){var r=n("constants"),o=process.cwd,i=null,s=process.env.GRACEFUL_FS_PLATFORM||process.platform;process.cwd=function(){return i||(i=o.call(process)),i};try{process.cwd()}catch(e){}var a=process.chdir;process.chdir=function(e){i=null,a.call(process,e)},e.exports=function(e){r.hasOwnProperty("O_SYMLINK")&&process.version.match(/^v0\.6\.[0-2]|^v0\.5\./)&&function(e){e.lchmod=function(t,n,o){e.open(t,r.O_WRONLY|r.O_SYMLINK,n,function(t,r){t?o&&o(t):e.fchmod(r,n,function(t){e.close(r,function(e){o&&o(t||e)})})})},e.lchmodSync=function(t,n){var o,i=e.openSync(t,r.O_WRONLY|r.O_SYMLINK,n),s=!0;try{o=e.fchmodSync(i,n),s=!1}finally{if(s)try{e.closeSync(i)}catch(e){}else e.closeSync(i)}return o}}(e);e.lutimes||function(e){r.hasOwnProperty("O_SYMLINK")?(e.lutimes=function(t,n,o,i){e.open(t,r.O_SYMLINK,function(t,r){t?i&&i(t):e.futimes(r,n,o,function(t){e.close(r,function(e){i&&i(t||e)})})})},e.lutimesSync=function(t,n,o){var i,s=e.openSync(t,r.O_SYMLINK),a=!0;try{i=e.futimesSync(s,n,o),a=!1}finally{if(a)try{e.closeSync(s)}catch(e){}else e.closeSync(s)}return i}):(e.lutimes=function(e,t,n,r){r&&process.nextTick(r)},e.lutimesSync=function(){})}(e);e.chown=i(e.chown),e.fchown=i(e.fchown),e.lchown=i(e.lchown),e.chmod=n(e.chmod),e.fchmod=n(e.fchmod),e.lchmod=n(e.lchmod),e.chownSync=a(e.chownSync),e.fchownSync=a(e.fchownSync),e.lchownSync=a(e.lchownSync),e.chmodSync=o(e.chmodSync),e.fchmodSync=o(e.fchmodSync),e.lchmodSync=o(e.lchmodSync),e.stat=c(e.stat),e.fstat=c(e.fstat),e.lstat=c(e.lstat),e.statSync=u(e.statSync),e.fstatSync=u(e.fstatSync),e.lstatSync=u(e.lstatSync),e.lchmod||(e.lchmod=function(e,t,n){n&&process.nextTick(n)},e.lchmodSync=function(){});e.lchown||(e.lchown=function(e,t,n,r){r&&process.nextTick(r)},e.lchownSync=function(){});"win32"===s&&(e.rename=(t=e.rename,function(n,r,o){var i=Date.now(),s=0;t(n,r,function a(c){if(c&&("EACCES"===c.code||"EPERM"===c.code)&&Date.now()-i<6e4)return setTimeout(function(){e.stat(r,function(e,i){e&&"ENOENT"===e.code?t(n,r,a):o(c)})},s),void(s<100&&(s+=10));o&&o(c)})}));var t;function n(t){return t?function(n,r,o){return t.call(e,n,r,function(e){l(e)&&(e=null),o&&o.apply(this,arguments)})}:t}function o(t){return t?function(n,r){try{return t.call(e,n,r)}catch(e){if(!l(e))throw e}}:t}function i(t){return t?function(n,r,o,i){return t.call(e,n,r,o,function(e){l(e)&&(e=null),i&&i.apply(this,arguments)})}:t}function a(t){return t?function(n,r,o){try{return t.call(e,n,r,o)}catch(e){if(!l(e))throw e}}:t}function c(t){return t?function(n,r){return t.call(e,n,function(e,t){if(!t)return r.apply(this,arguments);t.uid<0&&(t.uid+=4294967296),t.gid<0&&(t.gid+=4294967296),r&&r.apply(this,arguments)})}:t}function u(t){return t?function(n){var r=t.call(e,n);return r.uid<0&&(r.uid+=4294967296),r.gid<0&&(r.gid+=4294967296),r}:t}function l(e){if(!e)return!0;if("ENOSYS"===e.code)return!0;var t=!process.getuid||0!==process.getuid();return!(!t||"EINVAL"!==e.code&&"EPERM"!==e.code)}e.read=(h=e.read,function(t,n,r,o,i,s){var a;if(s&&"function"==typeof s){var c=0;a=function(u,l,p){if(u&&"EAGAIN"===u.code&&c<10)return c++,h.call(e,t,n,r,o,i,a);s.apply(this,arguments)}}return h.call(e,t,n,r,o,i,a)}),e.readSync=(p=e.readSync,function(t,n,r,o,i){for(var s=0;;)try{return p.call(e,t,n,r,o,i)}catch(e){if("EAGAIN"===e.code&&s<10){s++;continue}throw e}});var p;var h}},"./node_modules/imurmurhash/imurmurhash.js":function(e,t,n){
+module.exports=function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}return n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)n.d(r,o,function(t){return e[t]}.bind(null,o));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s="./app/main.dev.js")}({"./app/about-window.js":function(e,t,n){const{BrowserWindow:r,remote:o,shell:i}=n("electron"),s=n("path").join(process.resourcesPath,"icon.ico");let a=null;e.exports=function(e){if(null!==a)return a.focus(),a;const t={width:400,height:400,useContentSize:!0,titleBarStyle:"hidden-inset",icon:s,parent:e};return(a=new(r||o.BrowserWindow)(t)).once("closed",()=>{a=null}),a.loadURL(`file://${__dirname}/about.html`),a.webContents.on("will-navigate",(e,t)=>{e.preventDefault(),i.openExternal(t)}),a.webContents.on("new-window",(e,t)=>{e.preventDefault(),i.openExternal(t)}),a.webContents.once("dom-ready",()=>{let e;e={icon_path:s,product_name:"BloomRPC",copyright:"LGPL-3.0-or-later",homepage:"https://github.com/uw-labs/bloomrpc",description:"GRPC GUI client",license:"LGPL-3.0-or-later",bug_report_url:"https://github.com/uw-labs/bloomrpc/issues",version:"1.3.0",use_version_info:!0},a.webContents.send("about-window:info",e)}),a.once("ready-to-show",()=>{a.show()}),a.setMenu(null),a.setMenuBarVisibility(!1),a}},"./app/main.dev.js":function(e,t,n){const{app:r,BrowserWindow:o}=n("electron"),i=n("./app/menu.js");let s=null;n("./node_modules/source-map-support/source-map-support.js").install();r.on("window-all-closed",()=>{"darwin"!==process.platform&&r.quit()}),r.on("ready",async()=>{(s=new o({show:!1,width:1324,height:728,backgroundColor:"#f0f2f5"})).loadURL(`file://${__dirname}/app.html`),s.once("ready-to-show",()=>{if(!s)throw new Error('"mainWindow" is not defined');setTimeout(function(){s.show(),s.focus()},150)}),s.on("closed",()=>{s=null}),new i(s).buildMenu()})},"./app/menu.js":function(e,t,n){const{app:r,Menu:o,shell:i}=n("electron"),s=n("./app/about-window.js");e.exports=class{constructor(e){this.mainWindow=e}buildMenu(){const e="darwin"===process.platform?this.buildDarwinTemplate():this.buildDefaultTemplate(),t=o.buildFromTemplate(e);return o.setApplicationMenu(t),t}setupDevelopmentEnvironment(){this.mainWindow.openDevTools(),this.mainWindow.webContents.on("context-menu",(e,t)=>{const{x:n,y:r}=t;o.buildFromTemplate([{label:"Inspect element",click:()=>{this.mainWindow.inspectElement(n,r)}}]).popup(this.mainWindow)})}buildDarwinTemplate(){return[{label:"BloomRPC",submenu:[{label:"About BloomRPC",selector:"orderFrontStandardAboutPanel:"},{type:"separator"},{label:"Services",submenu:[]},{type:"separator"},{label:"Hide BloomRPC",accelerator:"Command+H",selector:"hide:"},{label:"Hide Others",accelerator:"Command+Shift+H",selector:"hideOtherApplications:"},{label:"Show All",selector:"unhideAllApplications:"},{type:"separator"},{label:"Quit",accelerator:"Command+Q",click:()=>{r.quit()}}]},{label:"Edit",submenu:[{label:"Undo",accelerator:"Command+Z",selector:"undo:"},{label:"Redo",accelerator:"Shift+Command+Z",selector:"redo:"},{type:"separator"},{label:"Cut",accelerator:"Command+X",selector:"cut:"},{label:"Copy",accelerator:"Command+C",selector:"copy:"},{label:"Paste",accelerator:"Command+V",selector:"paste:"},{label:"Select All",accelerator:"Command+A",selector:"selectAll:"},{label:"Clear Storage",accelerator:"Ctrl+Command+D",click(){n("./app/storage/index.ts").clearAll()}}]},{label:"View",submenu:[{label:"Toggle Full Screen",accelerator:"Ctrl+Command+F",click:()=>{this.mainWindow.setFullScreen(!this.mainWindow.isFullScreen())}}]},{label:"Window",submenu:[{label:"Minimize",accelerator:"Command+M",selector:"performMiniaturize:"},{label:"Close",accelerator:"Command+W",selector:"performClose:"},{type:"separator"},{label:"Bring All to Front",selector:"arrangeInFront:"}]},{label:"Help",submenu:[{label:"Learn More",click(){i.openExternal("https://github.com/uw-labs/bloomrpc")}},{label:"Documentation",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/blob/master/README.md")}},{label:"Community Discussions",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/issues")}},{label:"Search Issues",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/issues")}}]}]}buildDefaultTemplate(){return[{label:"&File",submenu:[{label:"&Open",accelerator:"Ctrl+O"},{label:"&Close",accelerator:"Ctrl+W",click:()=>{this.mainWindow.close()}},{label:"&Quit",accelerator:"Ctrl+Q",click:r.quit}]},{label:"&View",submenu:[{label:"Toggle &Full Screen",accelerator:"F11",click:()=>{this.mainWindow.setFullScreen(!this.mainWindow.isFullScreen())}}]},{label:"Help",submenu:[{label:"Learn More",click(){i.openExternal("https://github.com/uw-labs/bloomrpc")}},{label:"Documentation",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/blob/master/README.md")}},{label:"Community Discussions",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/issues")}},{label:"Search Issues",click(){i.openExternal("https://github.com/uw-labs/bloomrpc/issues")}},{label:"About BloomRPC",click:()=>{s(this.mainWindow)}}]}]}}},"./app/storage/editor.ts":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});const r=new(n("./node_modules/electron-store/index.js"))({name:"editor"});t.EditorStore=r;const o={URL:"url",PROTOS:"protos",TABS:"tabs",REQUESTS:"requests",INTERACTIVE:"interactive",METADATA:"metadata"};t.storeUrl=function(e){r.set(o.URL,e)},t.getUrl=function(){return r.get(o.URL)},t.storeProtos=function(e){r.set(o.PROTOS,e.map(e=>e.proto.filePath))},t.getProtos=function(){return r.get(o.PROTOS)},t.storeTabs=function(e){r.set(o.TABS,{activeKey:e.activeKey,tabs:e.tabs.map(e=>({methodName:e.methodName,serviceName:e.service.serviceName,protoPath:e.service.proto.filePath}))})},t.getTabs=function(){return r.get(o.TABS)},t.storeRequestInfo=function({id:e,url:t,data:n,inputs:i,metadata:s,interactive:a,tlsCertificate:c}){const u={id:e,url:t,data:i||n,metadata:s,interactive:a,tlsCertificate:c,createdAt:(new Date).toISOString()},l=r.get("requests",[]).filter(t=>t.id!==e);r.set(o.REQUESTS,[...l,u])},t.storeMetadata=function(e){r.set(o.METADATA,e)},t.getMetadata=function(){return r.get(o.METADATA)},t.getRequestInfo=function(e){return r.get(o.REQUESTS,[]).find(t=>t.id===e)},t.deleteRequestInfo=function(e){const t=r.get(o.REQUESTS,[]).filter(t=>t.id!==e);r.set("requests",t)},t.clearEditor=function(){r.clear()}},"./app/storage/importPaths.ts":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});const r=new(n("./node_modules/electron-store/index.js"))({name:"importPaths"}),o={IMPORT_PATH:"paths"};t.storeImportPaths=function(e){r.set(o.IMPORT_PATH,e)},t.getImportPaths=function(){return r.get(o.IMPORT_PATH,[""])},t.clearImportPaths=function(){return r.clear()}},"./app/storage/index.ts":function(e,t,n){"use strict";function r(e){for(var n in e)t.hasOwnProperty(n)||(t[n]=e[n])}Object.defineProperty(t,"__esModule",{value:!0});const o=n("./app/storage/editor.ts"),i=n("./app/storage/importPaths.ts"),s=n("./app/storage/tls.ts");r(n("./app/storage/editor.ts")),r(n("./app/storage/importPaths.ts")),r(n("./app/storage/tls.ts")),t.clearAll=function(){o.clearEditor(),i.clearImportPaths(),s.clearTLS()}},"./app/storage/tls.ts":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});const r=new(n("./node_modules/electron-store/index.js"))({name:"tls"}),o={CERTIFICATES:"certificates"};t.storeTLSList=function(e){r.set(o.CERTIFICATES,e)},t.getTLSList=function(){return r.get(o.CERTIFICATES,[])},t.clearTLS=function(){return r.clear()}},"./node_modules/7zip/index.js":function(e,t,n){var r,o,i=n("path").resolve,s=n("./node_modules/7zip/package.json").bin;e.exports=(r=s,o=function(e){return i(__dirname,e)},Object.keys(r).reduce(function(e,t){return e[t]=o(r[t]),e},{}))},"./node_modules/7zip/package.json":function(e){e.exports={name:"7zip",version:"0.0.6",description:"7zip Windows Package via Node.js",keywords:["7z","7zip","7-zip","windows","install"],repository:"git@github.com:fritx/win-7zip.git",bin:{"7z":"7zip-lite/7z.exe"},main:"index.js",scripts:{test:"mocha"},license:"GNU LGPL"}},"./node_modules/balanced-match/index.js":function(e,t,n){"use strict";function r(e,t,n){e instanceof RegExp&&(e=o(e,n)),t instanceof RegExp&&(t=o(t,n));var r=i(e,t,n);return r&&{start:r[0],end:r[1],pre:n.slice(0,r[0]),body:n.slice(r[0]+e.length,r[1]),post:n.slice(r[1]+t.length)}}function o(e,t){var n=t.match(e);return n?n[0]:null}function i(e,t,n){var r,o,i,s,a,c=n.indexOf(e),u=n.indexOf(t,c+1),l=c;if(c>=0&&u>0){for(r=[],i=n.length;l>=0&&!a;)l==c?(r.push(l),c=n.indexOf(e,l+1)):1==r.length?a=[r.pop(),u]:((o=r.pop())<i&&(i=o,s=u),u=n.indexOf(t,l+1)),l=c<u&&c>=0?c:u;r.length&&(a=[i,s])}return a}e.exports=r,r.range=i},"./node_modules/brace-expansion/index.js":function(e,t,n){var r=n("./node_modules/concat-map/index.js"),o=n("./node_modules/balanced-match/index.js");e.exports=function(e){if(!e)return[];"{}"===e.substr(0,2)&&(e="\\{\\}"+e.substr(2));return function e(t,n){var i=[];var s=o("{","}",t);if(!s||/\$$/.test(s.pre))return[t];var c=/^-?\d+\.\.-?\d+(?:\.\.-?\d+)?$/.test(s.body);var u=/^[a-zA-Z]\.\.[a-zA-Z](?:\.\.-?\d+)?$/.test(s.body);var p=c||u;var g=s.body.indexOf(",")>=0;if(!p&&!g)return s.post.match(/,.*\}/)?(t=s.pre+"{"+s.body+a+s.post,e(t)):[t];var y;if(p)y=s.body.split(/\.\./);else if(1===(y=function e(t){if(!t)return[""];var n=[];var r=o("{","}",t);if(!r)return t.split(",");var i=r.pre;var s=r.body;var a=r.post;var c=i.split(",");c[c.length-1]+="{"+s+"}";var u=e(a);a.length&&(c[c.length-1]+=u.shift(),c.push.apply(c,u));n.push.apply(n,c);return n}(s.body)).length&&1===(y=e(y[0],!1).map(h)).length){var v=s.post.length?e(s.post,!1):[""];return v.map(function(e){return s.pre+y[0]+e})}var _=s.pre;var v=s.post.length?e(s.post,!1):[""];var b;if(p){var w=l(y[0]),E=l(y[1]),S=Math.max(y[0].length,y[1].length),j=3==y.length?Math.abs(l(y[2])):1,O=d,x=E<w;x&&(j*=-1,O=m);var A=y.some(f);b=[];for(var C=w;O(C,E);C+=j){var L;if(u)"\\"===(L=String.fromCharCode(C))&&(L="");else if(L=String(C),A){var k=S-L.length;if(k>0){var R=new Array(k+1).join("0");L=C<0?"-"+R+L.slice(1):R+L}}b.push(L)}}else b=r(y,function(t){return e(t,!1)});for(var M=0;M<b.length;M++)for(var T=0;T<v.length;T++){var P=_+b[M]+v[T];(!n||p||P)&&i.push(P)}return i}(function(e){return e.split("\\\\").join(i).split("\\{").join(s).split("\\}").join(a).split("\\,").join(c).split("\\.").join(u)}(e),!0).map(p)};var i="\0SLASH"+Math.random()+"\0",s="\0OPEN"+Math.random()+"\0",a="\0CLOSE"+Math.random()+"\0",c="\0COMMA"+Math.random()+"\0",u="\0PERIOD"+Math.random()+"\0";function l(e){return parseInt(e,10)==e?parseInt(e,10):e.charCodeAt(0)}function p(e){return e.split(i).join("\\").split(s).join("{").split(a).join("}").split(c).join(",").split(u).join(".")}function h(e){return"{"+e+"}"}function f(e){return/^-?0\d/.test(e)}function d(e,t){return e<=t}function m(e,t){return e>=t}},"./node_modules/buffer-from/index.js":function(e,t){var n=Object.prototype.toString,r="function"==typeof Buffer.alloc&&"function"==typeof Buffer.allocUnsafe&&"function"==typeof Buffer.from;e.exports=function(e,t,o){if("number"==typeof e)throw new TypeError('"value" argument must not be a number');return i=e,"ArrayBuffer"===n.call(i).slice(8,-1)?function(e,t,n){t>>>=0;var o=e.byteLength-t;if(o<0)throw new RangeError("'offset' is out of bounds");if(void 0===n)n=o;else if((n>>>=0)>o)throw new RangeError("'length' is out of bounds");return r?Buffer.from(e.slice(t,t+n)):new Buffer(new Uint8Array(e.slice(t,t+n)))}(e,t,o):"string"==typeof e?function(e,t){if("string"==typeof t&&""!==t||(t="utf8"),!Buffer.isEncoding(t))throw new TypeError('"encoding" must be a valid string encoding');return r?Buffer.from(e,t):new Buffer(e,t)}(e,t):r?Buffer.from(e):new Buffer(e);var i}},"./node_modules/concat-map/index.js":function(e,t){e.exports=function(e,t){for(var r=[],o=0;o<e.length;o++){var i=t(e[o],o);n(i)?r.push.apply(r,i):r.push(i)}return r};var n=Array.isArray||function(e){return"[object Array]"===Object.prototype.toString.call(e)}},"./node_modules/conf/index.js":function(e,t,n){"use strict";(function(e){const t=n("fs"),r=n("path"),o=n("crypto"),i=n("assert"),s=n("events"),a=n("./node_modules/dot-prop/index.js"),c=n("./node_modules/make-dir/index.js"),u=n("./node_modules/pkg-up/index.js"),l=n("./node_modules/env-paths/index.js"),p=n("./node_modules/write-file-atomic/index.js"),h=()=>Object.create(null);delete n.c[__filename];const f=r.dirname(e.parent&&e.parent.filename||".");e.exports=class{constructor(e){const n=u.sync(f);if(!(e=Object.assign({projectName:n&&JSON.parse(t.readFileSync(n,"utf8")).name},e)).projectName&&!e.cwd)throw new Error("Project name could not be inferred. Please specify the `projectName` option.");(e=Object.assign({configName:"config",fileExtension:"json",projectSuffix:"nodejs"},e)).cwd||(e.cwd=l(e.projectName,{suffix:e.projectSuffix}).config),this.events=new s,this.encryptionKey=e.encryptionKey;const o=e.fileExtension?`.${e.fileExtension}`:"";this.path=r.resolve(e.cwd,`${e.configName}${o}`);const a=this.store,c=Object.assign(h(),e.defaults,a);try{i.deepEqual(a,c)}catch(e){this.store=c}}get(e,t){return a.get(this.store,e,t)}set(e,t){if("string"!=typeof e&&"object"!=typeof e)throw new TypeError(`Expected \`key\` to be of type \`string\` or \`object\`, got ${typeof e}`);if("object"!=typeof e&&void 0===t)throw new TypeError("Use `delete()` to clear values");const{store:n}=this;if("object"==typeof e)for(const t of Object.keys(e))a.set(n,t,e[t]);else a.set(n,e,t);this.store=n}has(e){return a.has(this.store,e)}delete(e){const{store:t}=this;a.delete(t,e),this.store=t}clear(){this.store=h()}onDidChange(e,t){if("string"!=typeof e)throw new TypeError(`Expected \`key\` to be of type \`string\`, got ${typeof e}`);if("function"!=typeof t)throw new TypeError(`Expected \`callback\` to be of type \`function\`, got ${typeof t}`);let n=this.get(e);const r=()=>{const r=n,o=this.get(e);try{i.deepEqual(o,r)}catch(e){n=o,t.call(this,o,r)}};return this.events.on("change",r),()=>this.events.removeListener("change",r)}get size(){return Object.keys(this.store).length}get store(){try{let e=t.readFileSync(this.path,this.encryptionKey?null:"utf8");if(this.encryptionKey)try{const t=o.createDecipher("aes-256-cbc",this.encryptionKey);e=Buffer.concat([t.update(e),t.final()])}catch(e){}return Object.assign(h(),JSON.parse(e))}catch(e){if("ENOENT"===e.code)return c.sync(r.dirname(this.path)),h();if("SyntaxError"===e.name)return h();throw e}}set store(e){c.sync(r.dirname(this.path));let t=JSON.stringify(e,null,"\t");if(this.encryptionKey){const e=o.createCipher("aes-256-cbc",this.encryptionKey);t=Buffer.concat([e.update(Buffer.from(t)),e.final()])}p.sync(this.path,t),this.events.emit("change")}*[Symbol.iterator](){const{store:e}=this;for(const t of Object.keys(e))yield[t,e[t]]}}}).call(this,n("./node_modules/webpack/buildin/module.js")(e))},"./node_modules/cross-unzip/index.js":function(e,t,n){"use strict";var r=n("child_process").spawn,o=Array.prototype.slice,i="win32"===process.platform?function(e,t,r){s(n("./node_modules/7zip/index.js")["7z"],["x",e,"-y","-o"+t],r)}:function(e,t,n){s("unzip",["-o",e,"-d",t],n)};function s(e,t,n){var i,s;i=n,s=!1,n=function(){s||(s=!0,i.apply(this,o.call(arguments)))};var a=r(e,t,{stdio:"ignore"});a.on("error",function(e){n(e)}),a.on("exit",function(e){n(e?new Error("Exited with code "+e):null)})}i.unzip=i,e.exports=i},"./node_modules/dot-prop/index.js":function(e,t,n){"use strict";const r=n("./node_modules/is-obj/index.js");function o(e){const t=e.split("."),n=[];for(let e=0;e<t.length;e++){let r=t[e];for(;"\\"===r[r.length-1]&&void 0!==t[e+1];)r=r.slice(0,-1)+".",r+=t[++e];n.push(r)}return n}e.exports={get(e,t,n){if(!r(e)||"string"!=typeof t)return void 0===n?e:n;const i=o(t);for(let t=0;t<i.length;t++){if(!Object.prototype.propertyIsEnumerable.call(e,i[t]))return n;if(null==(e=e[i[t]])){if(t!==i.length-1)return n;break}}return e},set(e,t,n){if(!r(e)||"string"!=typeof t)return e;const i=e,s=o(t);for(let t=0;t<s.length;t++){const o=s[t];r(e[o])||(e[o]={}),t===s.length-1&&(e[o]=n),e=e[o]}return i},delete(e,t){if(!r(e)||"string"!=typeof t)return;const n=o(t);for(let t=0;t<n.length;t++){const o=n[t];if(t===n.length-1)return void delete e[o];if(e=e[o],!r(e))return}},has(e,t){if(!r(e)||"string"!=typeof t)return!1;const n=o(t);for(let t=0;t<n.length;t++){if(!r(e))return!1;if(!(n[t]in e))return!1;e=e[n[t]]}return!0}}},"./node_modules/electron-devtools-installer/dist/downloadChromeExtension.js":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});var r=c(n("fs")),o=c(n("path")),i=c(n("./node_modules/rimraf/rimraf.js")),s=c(n("./node_modules/cross-unzip/index.js")),a=n("./node_modules/electron-devtools-installer/dist/utils.js");function c(e){return e&&e.__esModule?e:{default:e}}t.default=function e(t,n){var c=arguments.length>2&&void 0!==arguments[2]?arguments[2]:5,u=(0,a.getPath)();r.default.existsSync(u)||r.default.mkdirSync(u);var l=o.default.resolve(u+"/"+t);return new Promise(function(u,p){if(!r.default.existsSync(l)||n){r.default.existsSync(l)&&i.default.sync(l);var h="https://clients2.google.com/service/update2/crx?response=redirect&x=id%3D"+t+"%26uc&prodversion=32",f=o.default.resolve(l+".crx");(0,a.downloadFile)(h,f).then(function(){(0,s.default)(f,l,function(e){if(e&&!r.default.existsSync(o.default.resolve(l,"manifest.json")))return p(e);(0,a.changePermissions)(l,755),u(l)})}).catch(function(r){if(console.log("Failed to fetch extension, trying "+(c-1)+" more times"),c<=1)return p(r);setTimeout(function(){e(t,n,c-1).then(u).catch(p)},200)})}else u(l)})}},"./node_modules/electron-devtools-installer/dist/index.js":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.MOBX_DEVTOOLS=t.APOLLO_DEVELOPER_TOOLS=t.CYCLEJS_DEVTOOL=t.REACT_PERF=t.REDUX_DEVTOOLS=t.VUEJS_DEVTOOLS=t.ANGULARJS_BATARANG=t.JQUERY_DEBUGGER=t.BACKBONE_DEBUGGER=t.REACT_DEVELOPER_TOOLS=t.EMBER_INSPECTOR=void 0;var r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},o=n("electron"),i=p(o),s=p(n("fs")),a=p(n("path")),c=p(n("./node_modules/semver/semver.js")),u=p(n("./node_modules/electron-devtools-installer/dist/downloadChromeExtension.js")),l=n("./node_modules/electron-devtools-installer/dist/utils.js");function p(e){return e&&e.__esModule?e:{default:e}}var h=(o.remote||i.default).BrowserWindow,f={},d=a.default.resolve((0,l.getPath)(),"IDMap.json");if(s.default.existsSync(d))try{f=JSON.parse(s.default.readFileSync(d,"utf8"))}catch(e){console.error("electron-devtools-installer: Invalid JSON present in the IDMap file")}t.default=function e(t){var n=arguments.length>1&&void 0!==arguments[1]&&arguments[1];if(Array.isArray(t))return Promise.all(t.map(function(t){return e(t,n)}));var o=void 0;if("object"===(void 0===t?"undefined":r(t))&&t.id){o=t.id;var i=process.versions.electron.split("-")[0];if(!c.default.satisfies(i,t.electron))return Promise.reject(new Error("Version of Electron: "+i+" does not match required range "+t.electron+" for extension "+o))}else{if("string"!=typeof t)return Promise.reject(new Error('Invalid extensionReference passed in: "'+t+'"'));o=t}var a=f[o],l=a&&h.getDevToolsExtensions&&h.getDevToolsExtensions()[a];return!n&&l?Promise.resolve(f[o]):(0,u.default)(o,n).then(function(e){l&&h.removeDevToolsExtension(a);var t,n,r,i=h.addDevToolsExtension(e);return s.default.writeFileSync(d,JSON.stringify(Object.assign(f,(r=i,(n=o)in(t={})?Object.defineProperty(t,n,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[n]=r,t)))),Promise.resolve(i)})};t.EMBER_INSPECTOR={id:"bmdblncegkenkacieihfhpjfppoconhi",electron:">=1.2.1"},t.REACT_DEVELOPER_TOOLS={id:"fmkadmapgofadopljbjfkapdkoienihi",electron:">=1.2.1"},t.BACKBONE_DEBUGGER={id:"bhljhndlimiafopmmhjlgfpnnchjjbhd",electron:">=1.2.1"},t.JQUERY_DEBUGGER={id:"dbhhnnnpaeobfddmlalhnehgclcmjimi",electron:">=1.2.1"},t.ANGULARJS_BATARANG={id:"ighdmehidhipcmcojjgiloacoafjmpfk",electron:">=1.2.1"},t.VUEJS_DEVTOOLS={id:"nhdogjmejiglipccpnnnanhbledajbpd",electron:">=1.2.1"},t.REDUX_DEVTOOLS={id:"lmhkpmbekcpmknklioeibfkpmmfibljd",electron:">=1.2.1"},t.REACT_PERF={id:"hacmcodfllhbnekmghgdlplbdnahmhmm",electron:">=1.2.6"},t.CYCLEJS_DEVTOOL={id:"dfgplfmhhmdekalbpejekgfegkonjpfp",electron:">=1.2.1"},t.APOLLO_DEVELOPER_TOOLS={id:"jdkknkkbebbapilgoeccciglkfbmbnfm",electron:">=1.2.1"},t.MOBX_DEVTOOLS={id:"pfgnfdagidkfgccljigdamigbcnndkod",electron:">=1.2.1"}},"./node_modules/electron-devtools-installer/dist/utils.js":function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0}),t.changePermissions=t.downloadFile=t.getPath=void 0;var r=n("electron"),o=c(r),i=c(n("fs")),s=c(n("path")),a=c(n("https"));function c(e){return e&&e.__esModule?e:{default:e}}t.getPath=function(){var e=(r.remote||o.default).app.getPath("userData");return s.default.resolve(e+"/extensions")};var u=(r.remote||o.default).net,l=u?u.request:a.default.get;t.downloadFile=function e(t,n){return new Promise(function(r,o){var s=l(t);s.on("response",function(t){if(t.statusCode>=300&&t.statusCode<400&&t.headers.location)return e(t.headers.location,n).then(r).catch(o);t.pipe(i.default.createWriteStream(n)).on("close",r)}),s.on("error",o),s.end()})},t.changePermissions=function e(t,n){i.default.readdirSync(t).forEach(function(r){var o=s.default.join(t,r);i.default.chmodSync(o,parseInt(n,8)),i.default.statSync(o).isDirectory()&&e(o,n)})}},"./node_modules/electron-store/index.js":function(e,t,n){"use strict";const r=n("path"),o=n("electron"),i=n("./node_modules/conf/index.js");e.exports=class extends i{constructor(e){const t=(o.app||o.remote.app).getPath("userData");(e=Object.assign({name:"config"},e)).cwd?e.cwd=r.isAbsolute(e.cwd)?e.cwd:r.join(t,e.cwd):e.cwd=t,e.configName=e.name,delete e.name,super(e)}openInEditor(){o.shell.openItem(this.path)}}},"./node_modules/env-paths/index.js":function(e,t,n){"use strict";const r=n("path"),o=n("os"),i=o.homedir(),s=o.tmpdir(),a=process.env;e.exports=((e,t)=>{if("string"!=typeof e)throw new TypeError(`Expected string, got ${typeof e}`);return(t=Object.assign({suffix:"nodejs"},t)).suffix&&(e+=`-${t.suffix}`),"darwin"===process.platform?(e=>{const t=r.join(i,"Library");return{data:r.join(t,"Application Support",e),config:r.join(t,"Preferences",e),cache:r.join(t,"Caches",e),log:r.join(t,"Logs",e),temp:r.join(s,e)}})(e):"win32"===process.platform?(e=>{const t=a.LOCALAPPDATA||r.join(i,"AppData","Local");return{data:r.join(t,e,"Data"),config:r.join(t,e,"Config"),cache:r.join(t,e,"Cache"),log:r.join(t,e,"Log"),temp:r.join(s,e)}})(e):(e=>{const t=r.basename(i);return{data:r.join(a.XDG_DATA_HOME||r.join(i,".local","share"),e),config:r.join(a.XDG_CONFIG_HOME||r.join(i,".config"),e),cache:r.join(a.XDG_CACHE_HOME||r.join(i,".cache"),e),log:r.join(a.XDG_STATE_HOME||r.join(i,".local","state"),e),temp:r.join(s,t,e)}})(e)})},"./node_modules/fs.realpath/index.js":function(e,t,n){e.exports=l,l.realpath=l,l.sync=p,l.realpathSync=p,l.monkeypatch=function(){r.realpath=l,r.realpathSync=p},l.unmonkeypatch=function(){r.realpath=o,r.realpathSync=i};var r=n("fs"),o=r.realpath,i=r.realpathSync,s=process.version,a=/^v[0-5]\./.test(s),c=n("./node_modules/fs.realpath/old.js");function u(e){return e&&"realpath"===e.syscall&&("ELOOP"===e.code||"ENOMEM"===e.code||"ENAMETOOLONG"===e.code)}function l(e,t,n){if(a)return o(e,t,n);"function"==typeof t&&(n=t,t=null),o(e,t,function(r,o){u(r)?c.realpath(e,t,n):n(r,o)})}function p(e,t){if(a)return i(e,t);try{return i(e,t)}catch(n){if(u(n))return c.realpathSync(e,t);throw n}}},"./node_modules/fs.realpath/old.js":function(e,t,n){var r=n("path"),o="win32"===process.platform,i=n("fs"),s=process.env.NODE_DEBUG&&/fs/.test(process.env.NODE_DEBUG);function a(e){return"function"==typeof e?e:function(){var e;if(s){var t=new Error;e=function(e){e&&(t.message=e.message,n(e=t))}}else e=n;return e;function n(e){if(e){if(process.throwDeprecation)throw e;if(!process.noDeprecation){var t="fs: missing callback "+(e.stack||e.message);process.traceDeprecation?console.trace(t):console.error(t)}}}}()}r.normalize;if(o)var c=/(.*?)(?:[\/\\]+|$)/g;else c=/(.*?)(?:[\/]+|$)/g;if(o)var u=/^(?:[a-zA-Z]:|[\\\/]{2}[^\\\/]+[\\\/][^\\\/]+)?[\\\/]*/;else u=/^[\/]*/;t.realpathSync=function(e,t){if(e=r.resolve(e),t&&Object.prototype.hasOwnProperty.call(t,e))return t[e];var n,s,a,l,p=e,h={},f={};function d(){var t=u.exec(e);n=t[0].length,s=t[0],a=t[0],l="",o&&!f[a]&&(i.lstatSync(a),f[a]=!0)}for(d();n<e.length;){c.lastIndex=n;var m=c.exec(e);if(l=s,s+=m[0],a=l+m[1],n=c.lastIndex,!(f[a]||t&&t[a]===a)){var g;if(t&&Object.prototype.hasOwnProperty.call(t,a))g=t[a];else{var y=i.lstatSync(a);if(!y.isSymbolicLink()){f[a]=!0,t&&(t[a]=a);continue}var v=null;if(!o){var _=y.dev.toString(32)+":"+y.ino.toString(32);h.hasOwnProperty(_)&&(v=h[_])}null===v&&(i.statSync(a),v=i.readlinkSync(a)),g=r.resolve(l,v),t&&(t[a]=g),o||(h[_]=v)}e=r.resolve(g,e.slice(n)),d()}}return t&&(t[p]=e),e},t.realpath=function(e,t,n){if("function"!=typeof n&&(n=a(t),t=null),e=r.resolve(e),t&&Object.prototype.hasOwnProperty.call(t,e))return process.nextTick(n.bind(null,null,t[e]));var s,l,p,h,f=e,d={},m={};function g(){var t=u.exec(e);s=t[0].length,l=t[0],p=t[0],h="",o&&!m[p]?i.lstat(p,function(e){if(e)return n(e);m[p]=!0,y()}):process.nextTick(y)}function y(){if(s>=e.length)return t&&(t[f]=e),n(null,e);c.lastIndex=s;var r=c.exec(e);return h=l,l+=r[0],p=h+r[1],s=c.lastIndex,m[p]||t&&t[p]===p?process.nextTick(y):t&&Object.prototype.hasOwnProperty.call(t,p)?b(t[p]):i.lstat(p,v)}function v(e,r){if(e)return n(e);if(!r.isSymbolicLink())return m[p]=!0,t&&(t[p]=p),process.nextTick(y);if(!o){var s=r.dev.toString(32)+":"+r.ino.toString(32);if(d.hasOwnProperty(s))return _(null,d[s],p)}i.stat(p,function(e){if(e)return n(e);i.readlink(p,function(e,t){o||(d[s]=t),_(e,t)})})}function _(e,o,i){if(e)return n(e);var s=r.resolve(h,o);t&&(t[i]=s),b(s)}function b(t){e=r.resolve(t,e.slice(s)),g()}g()}},"./node_modules/glob/common.js":function(e,t,n){function r(e,t){return Object.prototype.hasOwnProperty.call(e,t)}t.alphasort=u,t.alphasorti=c,t.setopts=function(e,t,n){n||(n={});if(n.matchBase&&-1===t.indexOf("/")){if(n.noglobstar)throw new Error("base matching requires globstar");t="**/"+t}e.silent=!!n.silent,e.pattern=t,e.strict=!1!==n.strict,e.realpath=!!n.realpath,e.realpathCache=n.realpathCache||Object.create(null),e.follow=!!n.follow,e.dot=!!n.dot,e.mark=!!n.mark,e.nodir=!!n.nodir,e.nodir&&(e.mark=!0);e.sync=!!n.sync,e.nounique=!!n.nounique,e.nonull=!!n.nonull,e.nosort=!!n.nosort,e.nocase=!!n.nocase,e.stat=!!n.stat,e.noprocess=!!n.noprocess,e.absolute=!!n.absolute,e.maxLength=n.maxLength||1/0,e.cache=n.cache||Object.create(null),e.statCache=n.statCache||Object.create(null),e.symlinks=n.symlinks||Object.create(null),function(e,t){e.ignore=t.ignore||[],Array.isArray(e.ignore)||(e.ignore=[e.ignore]);e.ignore.length&&(e.ignore=e.ignore.map(l))}(e,n),e.changedCwd=!1;var i=process.cwd();r(n,"cwd")?(e.cwd=o.resolve(n.cwd),e.changedCwd=e.cwd!==i):e.cwd=i;e.root=n.root||o.resolve(e.cwd,"/"),e.root=o.resolve(e.root),"win32"===process.platform&&(e.root=e.root.replace(/\\/g,"/"));e.cwdAbs=s(e.cwd)?e.cwd:p(e,e.cwd),"win32"===process.platform&&(e.cwdAbs=e.cwdAbs.replace(/\\/g,"/"));e.nomount=!!n.nomount,n.nonegate=!0,n.nocomment=!0,e.minimatch=new a(t,n),e.options=e.minimatch.options},t.ownProp=r,t.makeAbs=p,t.finish=function(e){for(var t=e.nounique,n=t?[]:Object.create(null),r=0,o=e.matches.length;r<o;r++){var i=e.matches[r];if(i&&0!==Object.keys(i).length){var s=Object.keys(i);t?n.push.apply(n,s):s.forEach(function(e){n[e]=!0})}else if(e.nonull){var a=e.minimatch.globSet[r];t?n.push(a):n[a]=!0}}t||(n=Object.keys(n));e.nosort||(n=n.sort(e.nocase?c:u));if(e.mark){for(var r=0;r<n.length;r++)n[r]=e._mark(n[r]);e.nodir&&(n=n.filter(function(t){var n=!/\/$/.test(t),r=e.cache[t]||e.cache[p(e,t)];return n&&r&&(n="DIR"!==r&&!Array.isArray(r)),n}))}e.ignore.length&&(n=n.filter(function(t){return!h(e,t)}));e.found=n},t.mark=function(e,t){var n=p(e,t),r=e.cache[n],o=t;if(r){var i="DIR"===r||Array.isArray(r),s="/"===t.slice(-1);if(i&&!s?o+="/":!i&&s&&(o=o.slice(0,-1)),o!==t){var a=p(e,o);e.statCache[a]=e.statCache[n],e.cache[a]=e.cache[n]}}return o},t.isIgnored=h,t.childrenIgnored=function(e,t){return!!e.ignore.length&&e.ignore.some(function(e){return!(!e.gmatcher||!e.gmatcher.match(t))})};var o=n("path"),i=n("./node_modules/minimatch/minimatch.js"),s=n("./node_modules/path-is-absolute/index.js"),a=i.Minimatch;function c(e,t){return e.toLowerCase().localeCompare(t.toLowerCase())}function u(e,t){return e.localeCompare(t)}function l(e){var t=null;if("/**"===e.slice(-3)){var n=e.replace(/(\/\*\*)+$/,"");t=new a(n,{dot:!0})}return{matcher:new a(e,{dot:!0}),gmatcher:t}}function p(e,t){var n=t;return n="/"===t.charAt(0)?o.join(e.root,t):s(t)||""===t?t:e.changedCwd?o.resolve(e.cwd,t):o.resolve(t),"win32"===process.platform&&(n=n.replace(/\\/g,"/")),n}function h(e,t){return!!e.ignore.length&&e.ignore.some(function(e){return e.matcher.match(t)||!(!e.gmatcher||!e.gmatcher.match(t))})}},"./node_modules/glob/glob.js":function(e,t,n){e.exports=_;var r=n("fs"),o=n("./node_modules/fs.realpath/index.js"),i=n("./node_modules/minimatch/minimatch.js"),s=(i.Minimatch,n("./node_modules/inherits/inherits.js")),a=n("events").EventEmitter,c=n("path"),u=n("assert"),l=n("./node_modules/path-is-absolute/index.js"),p=n("./node_modules/glob/sync.js"),h=n("./node_modules/glob/common.js"),f=(h.alphasort,h.alphasorti,h.setopts),d=h.ownProp,m=n("./node_modules/inflight/inflight.js"),g=(n("util"),h.childrenIgnored),y=h.isIgnored,v=n("./node_modules/once/once.js");function _(e,t,n){if("function"==typeof t&&(n=t,t={}),t||(t={}),t.sync){if(n)throw new TypeError("callback provided to sync glob");return p(e,t)}return new w(e,t,n)}_.sync=p;var b=_.GlobSync=p.GlobSync;function w(e,t,n){if("function"==typeof t&&(n=t,t=null),t&&t.sync){if(n)throw new TypeError("callback provided to sync glob");return new b(e,t)}if(!(this instanceof w))return new w(e,t,n);f(this,e,t),this._didRealPath=!1;var r=this.minimatch.set.length;this.matches=new Array(r),"function"==typeof n&&(n=v(n),this.on("error",n),this.on("end",function(e){n(null,e)}));var o=this;if(this._processing=0,this._emitQueue=[],this._processQueue=[],this.paused=!1,this.noprocess)return this;if(0===r)return a();for(var i=!0,s=0;s<r;s++)this._process(this.minimatch.set[s],s,!1,a);function a(){--o._processing,o._processing<=0&&(i?process.nextTick(function(){o._finish()}):o._finish())}i=!1}_.glob=_,_.hasMagic=function(e,t){var n=function(e,t){if(null===t||"object"!=typeof t)return e;for(var n=Object.keys(t),r=n.length;r--;)e[n[r]]=t[n[r]];return e}({},t);n.noprocess=!0;var r=new w(e,n).minimatch.set;if(!e)return!1;if(r.length>1)return!0;for(var o=0;o<r[0].length;o++)if("string"!=typeof r[0][o])return!0;return!1},_.Glob=w,s(w,a),w.prototype._finish=function(){if(u(this instanceof w),!this.aborted){if(this.realpath&&!this._didRealpath)return this._realpath();h.finish(this),this.emit("end",this.found)}},w.prototype._realpath=function(){if(!this._didRealpath){this._didRealpath=!0;var e=this.matches.length;if(0===e)return this._finish();for(var t=this,n=0;n<this.matches.length;n++)this._realpathSet(n,r)}function r(){0==--e&&t._finish()}},w.prototype._realpathSet=function(e,t){var n=this.matches[e];if(!n)return t();var r=Object.keys(n),i=this,s=r.length;if(0===s)return t();var a=this.matches[e]=Object.create(null);r.forEach(function(n,r){n=i._makeAbs(n),o.realpath(n,i.realpathCache,function(r,o){r?"stat"===r.syscall?a[n]=!0:i.emit("error",r):a[o]=!0,0==--s&&(i.matches[e]=a,t())})})},w.prototype._mark=function(e){return h.mark(this,e)},w.prototype._makeAbs=function(e){return h.makeAbs(this,e)},w.prototype.abort=function(){this.aborted=!0,this.emit("abort")},w.prototype.pause=function(){this.paused||(this.paused=!0,this.emit("pause"))},w.prototype.resume=function(){if(this.paused){if(this.emit("resume"),this.paused=!1,this._emitQueue.length){var e=this._emitQueue.slice(0);this._emitQueue.length=0;for(var t=0;t<e.length;t++){var n=e[t];this._emitMatch(n[0],n[1])}}if(this._processQueue.length){var r=this._processQueue.slice(0);this._processQueue.length=0;for(t=0;t<r.length;t++){var o=r[t];this._processing--,this._process(o[0],o[1],o[2],o[3])}}}},w.prototype._process=function(e,t,n,r){if(u(this instanceof w),u("function"==typeof r),!this.aborted)if(this._processing++,this.paused)this._processQueue.push([e,t,n,r]);else{for(var o,s=0;"string"==typeof e[s];)s++;switch(s){case e.length:return void this._processSimple(e.join("/"),t,r);case 0:o=null;break;default:o=e.slice(0,s).join("/")}var a,c=e.slice(s);null===o?a=".":l(o)||l(e.join("/"))?(o&&l(o)||(o="/"+o),a=o):a=o;var p=this._makeAbs(a);if(g(this,a))return r();c[0]===i.GLOBSTAR?this._processGlobStar(o,a,p,c,t,n,r):this._processReaddir(o,a,p,c,t,n,r)}},w.prototype._processReaddir=function(e,t,n,r,o,i,s){var a=this;this._readdir(n,i,function(c,u){return a._processReaddir2(e,t,n,r,o,i,u,s)})},w.prototype._processReaddir2=function(e,t,n,r,o,i,s,a){if(!s)return a();for(var u=r[0],l=!!this.minimatch.negate,p=u._glob,h=this.dot||"."===p.charAt(0),f=[],d=0;d<s.length;d++){if("."!==(g=s[d]).charAt(0)||h)(l&&!e?!g.match(u):g.match(u))&&f.push(g)}var m=f.length;if(0===m)return a();if(1===r.length&&!this.mark&&!this.stat){this.matches[o]||(this.matches[o]=Object.create(null));for(d=0;d<m;d++){var g=f[d];e&&(g="/"!==e?e+"/"+g:e+g),"/"!==g.charAt(0)||this.nomount||(g=c.join(this.root,g)),this._emitMatch(o,g)}return a()}r.shift();for(d=0;d<m;d++){g=f[d];e&&(g="/"!==e?e+"/"+g:e+g),this._process([g].concat(r),o,i,a)}a()},w.prototype._emitMatch=function(e,t){if(!this.aborted&&!y(this,t))if(this.paused)this._emitQueue.push([e,t]);else{var n=l(t)?t:this._makeAbs(t);if(this.mark&&(t=this._mark(t)),this.absolute&&(t=n),!this.matches[e][t]){if(this.nodir){var r=this.cache[n];if("DIR"===r||Array.isArray(r))return}this.matches[e][t]=!0;var o=this.statCache[n];o&&this.emit("stat",t,o),this.emit("match",t)}}},w.prototype._readdirInGlobStar=function(e,t){if(!this.aborted){if(this.follow)return this._readdir(e,!1,t);var n=this,o=m("lstat\0"+e,function(r,o){if(r&&"ENOENT"===r.code)return t();var i=o&&o.isSymbolicLink();n.symlinks[e]=i,i||!o||o.isDirectory()?n._readdir(e,!1,t):(n.cache[e]="FILE",t())});o&&r.lstat(e,o)}},w.prototype._readdir=function(e,t,n){if(!this.aborted&&(n=m("readdir\0"+e+"\0"+t,n))){if(t&&!d(this.symlinks,e))return this._readdirInGlobStar(e,n);if(d(this.cache,e)){var o=this.cache[e];if(!o||"FILE"===o)return n();if(Array.isArray(o))return n(null,o)}r.readdir(e,function(e,t,n){return function(r,o){r?e._readdirError(t,r,n):e._readdirEntries(t,o,n)}}(this,e,n))}},w.prototype._readdirEntries=function(e,t,n){if(!this.aborted){if(!this.mark&&!this.stat)for(var r=0;r<t.length;r++){var o=t[r];o="/"===e?e+o:e+"/"+o,this.cache[o]=!0}return this.cache[e]=t,n(null,t)}},w.prototype._readdirError=function(e,t,n){if(!this.aborted){switch(t.code){case"ENOTSUP":case"ENOTDIR":var r=this._makeAbs(e);if(this.cache[r]="FILE",r===this.cwdAbs){var o=new Error(t.code+" invalid cwd "+this.cwd);o.path=this.cwd,o.code=t.code,this.emit("error",o),this.abort()}break;case"ENOENT":case"ELOOP":case"ENAMETOOLONG":case"UNKNOWN":this.cache[this._makeAbs(e)]=!1;break;default:this.cache[this._makeAbs(e)]=!1,this.strict&&(this.emit("error",t),this.abort()),this.silent||console.error("glob error",t)}return n()}},w.prototype._processGlobStar=function(e,t,n,r,o,i,s){var a=this;this._readdir(n,i,function(c,u){a._processGlobStar2(e,t,n,r,o,i,u,s)})},w.prototype._processGlobStar2=function(e,t,n,r,o,i,s,a){if(!s)return a();var c=r.slice(1),u=e?[e]:[],l=u.concat(c);this._process(l,o,!1,a);var p=this.symlinks[n],h=s.length;if(p&&i)return a();for(var f=0;f<h;f++){if("."!==s[f].charAt(0)||this.dot){var d=u.concat(s[f],c);this._process(d,o,!0,a);var m=u.concat(s[f],r);this._process(m,o,!0,a)}}a()},w.prototype._processSimple=function(e,t,n){var r=this;this._stat(e,function(o,i){r._processSimple2(e,t,o,i,n)})},w.prototype._processSimple2=function(e,t,n,r,o){if(this.matches[t]||(this.matches[t]=Object.create(null)),!r)return o();if(e&&l(e)&&!this.nomount){var i=/[\/\\]$/.test(e);"/"===e.charAt(0)?e=c.join(this.root,e):(e=c.resolve(this.root,e),i&&(e+="/"))}"win32"===process.platform&&(e=e.replace(/\\/g,"/")),this._emitMatch(t,e),o()},w.prototype._stat=function(e,t){var n=this._makeAbs(e),o="/"===e.slice(-1);if(e.length>this.maxLength)return t();if(!this.stat&&d(this.cache,n)){var i=this.cache[n];if(Array.isArray(i)&&(i="DIR"),!o||"DIR"===i)return t(null,i);if(o&&"FILE"===i)return t()}var s=this.statCache[n];if(void 0!==s){if(!1===s)return t(null,s);var a=s.isDirectory()?"DIR":"FILE";return o&&"FILE"===a?t():t(null,a,s)}var c=this,u=m("stat\0"+n,function(o,i){if(i&&i.isSymbolicLink())return r.stat(n,function(r,o){r?c._stat2(e,n,null,i,t):c._stat2(e,n,r,o,t)});c._stat2(e,n,o,i,t)});u&&r.lstat(n,u)},w.prototype._stat2=function(e,t,n,r,o){if(n&&("ENOENT"===n.code||"ENOTDIR"===n.code))return this.statCache[t]=!1,o();var i="/"===e.slice(-1);if(this.statCache[t]=r,"/"===t.slice(-1)&&r&&!r.isDirectory())return o(null,!1,r);var s=!0;return r&&(s=r.isDirectory()?"DIR":"FILE"),this.cache[t]=this.cache[t]||s,i&&"FILE"===s?o():o(null,s,r)}},"./node_modules/glob/sync.js":function(e,t,n){e.exports=d,d.GlobSync=m;var r=n("fs"),o=n("./node_modules/fs.realpath/index.js"),i=n("./node_modules/minimatch/minimatch.js"),s=(i.Minimatch,n("./node_modules/glob/glob.js").Glob,n("util"),n("path")),a=n("assert"),c=n("./node_modules/path-is-absolute/index.js"),u=n("./node_modules/glob/common.js"),l=(u.alphasort,u.alphasorti,u.setopts),p=u.ownProp,h=u.childrenIgnored,f=u.isIgnored;function d(e,t){if("function"==typeof t||3===arguments.length)throw new TypeError("callback provided to sync glob\nSee: https://github.com/isaacs/node-glob/issues/167");return new m(e,t).found}function m(e,t){if(!e)throw new Error("must provide pattern");if("function"==typeof t||3===arguments.length)throw new TypeError("callback provided to sync glob\nSee: https://github.com/isaacs/node-glob/issues/167");if(!(this instanceof m))return new m(e,t);if(l(this,e,t),this.noprocess)return this;var n=this.minimatch.set.length;this.matches=new Array(n);for(var r=0;r<n;r++)this._process(this.minimatch.set[r],r,!1);this._finish()}m.prototype._finish=function(){if(a(this instanceof m),this.realpath){var e=this;this.matches.forEach(function(t,n){var r=e.matches[n]=Object.create(null);for(var i in t)try{i=e._makeAbs(i),r[o.realpathSync(i,e.realpathCache)]=!0}catch(t){if("stat"!==t.syscall)throw t;r[e._makeAbs(i)]=!0}})}u.finish(this)},m.prototype._process=function(e,t,n){a(this instanceof m);for(var r,o=0;"string"==typeof e[o];)o++;switch(o){case e.length:return void this._processSimple(e.join("/"),t);case 0:r=null;break;default:r=e.slice(0,o).join("/")}var s,u=e.slice(o);null===r?s=".":c(r)||c(e.join("/"))?(r&&c(r)||(r="/"+r),s=r):s=r;var l=this._makeAbs(s);h(this,s)||(u[0]===i.GLOBSTAR?this._processGlobStar(r,s,l,u,t,n):this._processReaddir(r,s,l,u,t,n))},m.prototype._processReaddir=function(e,t,n,r,o,i){var a=this._readdir(n,i);if(a){for(var c=r[0],u=!!this.minimatch.negate,l=c._glob,p=this.dot||"."===l.charAt(0),h=[],f=0;f<a.length;f++){if("."!==(g=a[f]).charAt(0)||p)(u&&!e?!g.match(c):g.match(c))&&h.push(g)}var d=h.length;if(0!==d)if(1!==r.length||this.mark||this.stat){r.shift();for(f=0;f<d;f++){var m;g=h[f];m=e?[e,g]:[g],this._process(m.concat(r),o,i)}}else{this.matches[o]||(this.matches[o]=Object.create(null));for(var f=0;f<d;f++){var g=h[f];e&&(g="/"!==e.slice(-1)?e+"/"+g:e+g),"/"!==g.charAt(0)||this.nomount||(g=s.join(this.root,g)),this._emitMatch(o,g)}}}},m.prototype._emitMatch=function(e,t){if(!f(this,t)){var n=this._makeAbs(t);if(this.mark&&(t=this._mark(t)),this.absolute&&(t=n),!this.matches[e][t]){if(this.nodir){var r=this.cache[n];if("DIR"===r||Array.isArray(r))return}this.matches[e][t]=!0,this.stat&&this._stat(t)}}},m.prototype._readdirInGlobStar=function(e){if(this.follow)return this._readdir(e,!1);var t,n;try{n=r.lstatSync(e)}catch(e){if("ENOENT"===e.code)return null}var o=n&&n.isSymbolicLink();return this.symlinks[e]=o,o||!n||n.isDirectory()?t=this._readdir(e,!1):this.cache[e]="FILE",t},m.prototype._readdir=function(e,t){if(t&&!p(this.symlinks,e))return this._readdirInGlobStar(e);if(p(this.cache,e)){var n=this.cache[e];if(!n||"FILE"===n)return null;if(Array.isArray(n))return n}try{return this._readdirEntries(e,r.readdirSync(e))}catch(t){return this._readdirError(e,t),null}},m.prototype._readdirEntries=function(e,t){if(!this.mark&&!this.stat)for(var n=0;n<t.length;n++){var r=t[n];r="/"===e?e+r:e+"/"+r,this.cache[r]=!0}return this.cache[e]=t,t},m.prototype._readdirError=function(e,t){switch(t.code){case"ENOTSUP":case"ENOTDIR":var n=this._makeAbs(e);if(this.cache[n]="FILE",n===this.cwdAbs){var r=new Error(t.code+" invalid cwd "+this.cwd);throw r.path=this.cwd,r.code=t.code,r}break;case"ENOENT":case"ELOOP":case"ENAMETOOLONG":case"UNKNOWN":this.cache[this._makeAbs(e)]=!1;break;default:if(this.cache[this._makeAbs(e)]=!1,this.strict)throw t;this.silent||console.error("glob error",t)}},m.prototype._processGlobStar=function(e,t,n,r,o,i){var s=this._readdir(n,i);if(s){var a=r.slice(1),c=e?[e]:[],u=c.concat(a);this._process(u,o,!1);var l=s.length;if(!this.symlinks[n]||!i)for(var p=0;p<l;p++){if("."!==s[p].charAt(0)||this.dot){var h=c.concat(s[p],a);this._process(h,o,!0);var f=c.concat(s[p],r);this._process(f,o,!0)}}}},m.prototype._processSimple=function(e,t){var n=this._stat(e);if(this.matches[t]||(this.matches[t]=Object.create(null)),n){if(e&&c(e)&&!this.nomount){var r=/[\/\\]$/.test(e);"/"===e.charAt(0)?e=s.join(this.root,e):(e=s.resolve(this.root,e),r&&(e+="/"))}"win32"===process.platform&&(e=e.replace(/\\/g,"/")),this._emitMatch(t,e)}},m.prototype._stat=function(e){var t=this._makeAbs(e),n="/"===e.slice(-1);if(e.length>this.maxLength)return!1;if(!this.stat&&p(this.cache,t)){var o=this.cache[t];if(Array.isArray(o)&&(o="DIR"),!n||"DIR"===o)return o;if(n&&"FILE"===o)return!1}var i=this.statCache[t];if(!i){var s;try{s=r.lstatSync(t)}catch(e){if(e&&("ENOENT"===e.code||"ENOTDIR"===e.code))return this.statCache[t]=!1,!1}if(s&&s.isSymbolicLink())try{i=r.statSync(t)}catch(e){i=s}else i=s}this.statCache[t]=i;o=!0;return i&&(o=i.isDirectory()?"DIR":"FILE"),this.cache[t]=this.cache[t]||o,(!n||"FILE"!==o)&&o},m.prototype._mark=function(e){return u.mark(this,e)},m.prototype._makeAbs=function(e){return u.makeAbs(this,e)}},"./node_modules/graceful-fs/clone.js":function(e,t,n){"use strict";e.exports=function(e){if(null===e||"object"!=typeof e)return e;if(e instanceof Object)var t={__proto__:e.__proto__};else var t=Object.create(null);return Object.getOwnPropertyNames(e).forEach(function(n){Object.defineProperty(t,n,Object.getOwnPropertyDescriptor(e,n))}),t}},"./node_modules/graceful-fs/graceful-fs.js":function(e,t,n){var r=n("fs"),o=n("./node_modules/graceful-fs/polyfills.js"),i=n("./node_modules/graceful-fs/legacy-streams.js"),s=n("./node_modules/graceful-fs/clone.js"),a=[],c=n("util");var u,l,p=function(){};function h(e){o(e),e.gracefulify=h,e.FileReadStream=p,e.FileWriteStream=m,e.createReadStream=function(e,t){return new p(e,t)},e.createWriteStream=function(e,t){return new m(e,t)};var t=e.readFile;e.readFile=function(e,n,r){"function"==typeof n&&(r=n,n=null);return function e(n,r,o){return t(n,r,function(t){!t||"EMFILE"!==t.code&&"ENFILE"!==t.code?("function"==typeof o&&o.apply(this,arguments),d()):f([e,[n,r,o]])})}(e,n,r)};var n=e.writeFile;e.writeFile=function(e,t,r,o){"function"==typeof r&&(o=r,r=null);return function e(t,r,o,i){return n(t,r,o,function(n){!n||"EMFILE"!==n.code&&"ENFILE"!==n.code?("function"==typeof i&&i.apply(this,arguments),d()):f([e,[t,r,o,i]])})}(e,t,r,o)};var r=e.appendFile;r&&(e.appendFile=function(e,t,n,o){"function"==typeof n&&(o=n,n=null);return function e(t,n,o,i){return r(t,n,o,function(r){!r||"EMFILE"!==r.code&&"ENFILE"!==r.code?("function"==typeof i&&i.apply(this,arguments),d()):f([e,[t,n,o,i]])})}(e,t,n,o)});var s=e.readdir;function a(t){return s.apply(e,t)}if(e.readdir=function(e,t,n){var r=[e];"function"!=typeof t?r.push(t):n=t;return r.push(function(e,t){t&&t.sort&&t.sort(),!e||"EMFILE"!==e.code&&"ENFILE"!==e.code?("function"==typeof n&&n.apply(this,arguments),d()):f([a,[r]])}),a(r)},"v0.8"===process.version.substr(0,4)){var c=i(e);p=c.ReadStream,m=c.WriteStream}var u=e.ReadStream;u&&(p.prototype=Object.create(u.prototype),p.prototype.open=function(){var e=this;y(e.path,e.flags,e.mode,function(t,n){t?(e.autoClose&&e.destroy(),e.emit("error",t)):(e.fd=n,e.emit("open",n),e.read())})});var l=e.WriteStream;function p(e,t){return this instanceof p?(u.apply(this,arguments),this):p.apply(Object.create(p.prototype),arguments)}function m(e,t){return this instanceof m?(l.apply(this,arguments),this):m.apply(Object.create(m.prototype),arguments)}l&&(m.prototype=Object.create(l.prototype),m.prototype.open=function(){var e=this;y(e.path,e.flags,e.mode,function(t,n){t?(e.destroy(),e.emit("error",t)):(e.fd=n,e.emit("open",n))})}),e.ReadStream=p,e.WriteStream=m;var g=e.open;function y(e,t,n,r){return"function"==typeof n&&(r=n,n=null),function e(t,n,r,o){return g(t,n,r,function(i,s){!i||"EMFILE"!==i.code&&"ENFILE"!==i.code?("function"==typeof o&&o.apply(this,arguments),d()):f([e,[t,n,r,o]])})}(e,t,n,r)}return e.open=y,e}function f(e){p("ENQUEUE",e[0].name,e[1]),a.push(e)}function d(){var e=a.shift();e&&(p("RETRY",e[0].name,e[1]),e[0].apply(null,e[1]))}c.debuglog?p=c.debuglog("gfs4"):/\bgfs4\b/i.test(process.env.NODE_DEBUG||"")&&(p=function(){var e=c.format.apply(c,arguments);e="GFS4: "+e.split(/\n/).join("\nGFS4: "),console.error(e)}),/\bgfs4\b/i.test(process.env.NODE_DEBUG||"")&&process.on("exit",function(){p(a),n("assert").equal(a.length,0)}),e.exports=h(s(r)),process.env.TEST_GRACEFUL_FS_GLOBAL_PATCH&&!r.__patched&&(e.exports=h(r),r.__patched=!0),e.exports.close=(u=r.close,function(e,t){return u.call(r,e,function(e){e||d(),"function"==typeof t&&t.apply(this,arguments)})}),e.exports.closeSync=(l=r.closeSync,function(e){var t=l.apply(r,arguments);return d(),t}),/\bgraceful-fs\b/.test(r.closeSync.toString())||(r.closeSync=e.exports.closeSync,r.close=e.exports.close)},"./node_modules/graceful-fs/legacy-streams.js":function(e,t,n){var r=n("stream").Stream;e.exports=function(e){return{ReadStream:function t(n,o){if(!(this instanceof t))return new t(n,o);r.call(this);var i=this;this.path=n;this.fd=null;this.readable=!0;this.paused=!1;this.flags="r";this.mode=438;this.bufferSize=65536;o=o||{};var s=Object.keys(o);for(var a=0,c=s.length;a<c;a++){var u=s[a];this[u]=o[u]}this.encoding&&this.setEncoding(this.encoding);if(void 0!==this.start){if("number"!=typeof this.start)throw TypeError("start must be a Number");if(void 0===this.end)this.end=1/0;else if("number"!=typeof this.end)throw TypeError("end must be a Number");if(this.start>this.end)throw new Error("start must be <= end");this.pos=this.start}if(null!==this.fd)return void process.nextTick(function(){i._read()});e.open(this.path,this.flags,this.mode,function(e,t){if(e)return i.emit("error",e),void(i.readable=!1);i.fd=t,i.emit("open",t),i._read()})},WriteStream:function t(n,o){if(!(this instanceof t))return new t(n,o);r.call(this);this.path=n;this.fd=null;this.writable=!0;this.flags="w";this.encoding="binary";this.mode=438;this.bytesWritten=0;o=o||{};var i=Object.keys(o);for(var s=0,a=i.length;s<a;s++){var c=i[s];this[c]=o[c]}if(void 0!==this.start){if("number"!=typeof this.start)throw TypeError("start must be a Number");if(this.start<0)throw new Error("start must be >= zero");this.pos=this.start}this.busy=!1;this._queue=[];null===this.fd&&(this._open=e.open,this._queue.push([this._open,this.path,this.flags,this.mode,void 0]),this.flush())}}}},"./node_modules/graceful-fs/polyfills.js":function(e,t,n){var r=n("constants"),o=process.cwd,i=null,s=process.env.GRACEFUL_FS_PLATFORM||process.platform;process.cwd=function(){return i||(i=o.call(process)),i};try{process.cwd()}catch(e){}var a=process.chdir;process.chdir=function(e){i=null,a.call(process,e)},e.exports=function(e){r.hasOwnProperty("O_SYMLINK")&&process.version.match(/^v0\.6\.[0-2]|^v0\.5\./)&&function(e){e.lchmod=function(t,n,o){e.open(t,r.O_WRONLY|r.O_SYMLINK,n,function(t,r){t?o&&o(t):e.fchmod(r,n,function(t){e.close(r,function(e){o&&o(t||e)})})})},e.lchmodSync=function(t,n){var o,i=e.openSync(t,r.O_WRONLY|r.O_SYMLINK,n),s=!0;try{o=e.fchmodSync(i,n),s=!1}finally{if(s)try{e.closeSync(i)}catch(e){}else e.closeSync(i)}return o}}(e);e.lutimes||function(e){r.hasOwnProperty("O_SYMLINK")?(e.lutimes=function(t,n,o,i){e.open(t,r.O_SYMLINK,function(t,r){t?i&&i(t):e.futimes(r,n,o,function(t){e.close(r,function(e){i&&i(t||e)})})})},e.lutimesSync=function(t,n,o){var i,s=e.openSync(t,r.O_SYMLINK),a=!0;try{i=e.futimesSync(s,n,o),a=!1}finally{if(a)try{e.closeSync(s)}catch(e){}else e.closeSync(s)}return i}):(e.lutimes=function(e,t,n,r){r&&process.nextTick(r)},e.lutimesSync=function(){})}(e);e.chown=i(e.chown),e.fchown=i(e.fchown),e.lchown=i(e.lchown),e.chmod=n(e.chmod),e.fchmod=n(e.fchmod),e.lchmod=n(e.lchmod),e.chownSync=a(e.chownSync),e.fchownSync=a(e.fchownSync),e.lchownSync=a(e.lchownSync),e.chmodSync=o(e.chmodSync),e.fchmodSync=o(e.fchmodSync),e.lchmodSync=o(e.lchmodSync),e.stat=c(e.stat),e.fstat=c(e.fstat),e.lstat=c(e.lstat),e.statSync=u(e.statSync),e.fstatSync=u(e.fstatSync),e.lstatSync=u(e.lstatSync),e.lchmod||(e.lchmod=function(e,t,n){n&&process.nextTick(n)},e.lchmodSync=function(){});e.lchown||(e.lchown=function(e,t,n,r){r&&process.nextTick(r)},e.lchownSync=function(){});"win32"===s&&(e.rename=(t=e.rename,function(n,r,o){var i=Date.now(),s=0;t(n,r,function a(c){if(c&&("EACCES"===c.code||"EPERM"===c.code)&&Date.now()-i<6e4)return setTimeout(function(){e.stat(r,function(e,i){e&&"ENOENT"===e.code?t(n,r,a):o(c)})},s),void(s<100&&(s+=10));o&&o(c)})}));var t;function n(t){return t?function(n,r,o){return t.call(e,n,r,function(e){l(e)&&(e=null),o&&o.apply(this,arguments)})}:t}function o(t){return t?function(n,r){try{return t.call(e,n,r)}catch(e){if(!l(e))throw e}}:t}function i(t){return t?function(n,r,o,i){return t.call(e,n,r,o,function(e){l(e)&&(e=null),i&&i.apply(this,arguments)})}:t}function a(t){return t?function(n,r,o){try{return t.call(e,n,r,o)}catch(e){if(!l(e))throw e}}:t}function c(t){return t?function(n,r){return t.call(e,n,function(e,t){if(!t)return r.apply(this,arguments);t.uid<0&&(t.uid+=4294967296),t.gid<0&&(t.gid+=4294967296),r&&r.apply(this,arguments)})}:t}function u(t){return t?function(n){var r=t.call(e,n);return r.uid<0&&(r.uid+=4294967296),r.gid<0&&(r.gid+=4294967296),r}:t}function l(e){if(!e)return!0;if("ENOSYS"===e.code)return!0;var t=!process.getuid||0!==process.getuid();return!(!t||"EINVAL"!==e.code&&"EPERM"!==e.code)}e.read=(h=e.read,function(t,n,r,o,i,s){var a;if(s&&"function"==typeof s){var c=0;a=function(u,l,p){if(u&&"EAGAIN"===u.code&&c<10)return c++,h.call(e,t,n,r,o,i,a);s.apply(this,arguments)}}return h.call(e,t,n,r,o,i,a)}),e.readSync=(p=e.readSync,function(t,n,r,o,i){for(var s=0;;)try{return p.call(e,t,n,r,o,i)}catch(e){if("EAGAIN"===e.code&&s<10){s++;continue}throw e}});var p;var h}},"./node_modules/imurmurhash/imurmurhash.js":function(e,t,n){
 /**
  * @preserve
  * JS Implementation of incremental MurmurHash3 (r150) (as of May 10, 2013)
